==> scripts/1-dc-data -curate-and-processing.rmd <==
---
# title: "Data Preprocessing"
output:
  html_document:
    df_print: paged
  pdf_document: default
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(dplyr)
library(plotrix)
library(plotly)
library(tools)
library(janitor)
```

```{r, echo=FALSE}
root_dir <- getwd()
project_directory<-dirname(root_dir)
data_dir <- file.path(project_directory, 'raw-data')
curated_data_dir <- file.path(project_directory, 'curated-data')

data_file_name<-'StandardPart_Data.csv'
# data_file_name<-'StandardPart_DATA_LABELS.csv'


Standard_data<-read.csv(file.path(data_dir,data_file_name),stringsAsFactors = FALSE)
Standard_data <- Standard_data[-c(1:80, 82:94), ]

Standard_data$email[1]="ljohnsson12@gmail.com"
Standard_data$confirm_email[1]="ljohnsson12@gmail.com"

missing_email<-Standard_data[c('111','120','129', '148', '154', '161', '186', '194', '233', '246','347', '413', '469', '510', '566', '578', '612', '618', '633' ),]


# Standard_data <- Standard_data[Standard_data$phone_a_friend !="",]

Standard_data<- subset(Standard_data, Standard_data$standard_part_complete==2)
Standard_data<-rbind(Standard_data, missing_email)

Standard_data <- Standard_data[-c(5:9, 12), ]
Standard_data_index <- Standard_data[, -c(2:5)]
old_name<-colnames(Standard_data_index)
# Standard_data <- Standard_data[-c(82:95), ]
```

```{r, echo=FALSE}
# data_file_name<-'StandardPart_Data.csv'
data_file_name<-'StandardPart_DATA_LABELS.csv'


Standard_data<-read.csv(file.path(data_dir,data_file_name),stringsAsFactors = FALSE)
Standard_data <- Standard_data[-c(1:80, 82:94), ]

Standard_data$Please.provide.your.email.address.The.system.will.use.this.email.address.to.send.you.an.acknowledgment..as.well.as.the.Amazon.gift.certificate..in.case.you.are.the.raffle.winner.[1]="ljohnsson12@gmail.com"
Standard_data$Please.confirm.your.email.address[1]="ljohnsson12@gmail.com"

missing_email<-Standard_data[c('111','120','129', '148', '154', '161', '186', '194', '233', '246','347', '413', '469', '510', '566', '578', '612', '618', '633' ),]

# Standard_data <- Standard_data[Standard_data$Phone.a.friend !="",]

Standard_data<- subset(Standard_data, Standard_data$Complete.=="Complete")
Standard_data<-rbind(Standard_data, missing_email)

Standard_data <- Standard_data[-c(5:9, 12), ]
Standard_data_label <- Standard_data[, -c(2:5)]
colnames(Standard_data_label)<- old_name
```

```{r, echo=FALSE}
# final_data<-Standard_data_label[, c(1:80)]
# final_data<-cbind(final_data, Standard_data_index[,c(81:135)])

final_data<-Standard_data_label[, c(1:44)]
final_data<-cbind(final_data, Standard_data_index[,c(45:49)])
final_data<-cbind(final_data, Standard_data_label[,c(50:69)])
final_data<-cbind(final_data, Standard_data_index[,c(70:135)])

for (i in 1:nrow(final_data)) {
  
  # if (final_data$departmental_affiliation[i] == "Other") {
  #   final_data$departmental_affiliation[i] = final_data$please_write_down_the_depa[i]
  # }
  if (final_data$how_many_months_of_your_sa[i] == "Other") {
    final_data$how_many_months_of_your_sa[i] = final_data$please_specify_the_months[i]
  }
  
  if (final_data$funding_agency[i] == "Other") {
    final_data$funding_agency[i] = final_data$specify_the_name[i]
  }
  
  if (final_data$where_you_typically_wor[i] == "Other") {
    final_data$where_you_typically_wor[i] = final_data$please_specify_your_work_p[i]
  }
  
  if (final_data$wha_average_number[i] == "â‰¥ 10") {
    final_data$wha_average_number[i] = ">= 10"
  }
  
  if (final_data$what_is_the_average_number[i] == "â‰¥ 10") {
    final_data$what_is_the_average_number[i] = ">= 10"
  }
  
  if (final_data$funding_agency[i] == "Education Advancement Foundation (a private foundation)" | final_data$funding_agency[i] =="Industry, private foundations") {
    final_data$funding_agency[i] = "Private Foundation"
  }
  if (final_data$funding_agency[i] == "trb"){
    final_data$funding_agency[i] = "TRB"
  }
  
  if (final_data$test[i]=="Hands-on [... plus experiments/analysis]"){
    final_data$test[i]="Hands-on"
  }else{
    final_data$test[i]="Hands-off"
  }
  
}

final_data$ti_research<-gsub("[^0-9.-]", "", final_data$ti_research)
final_data$ti_teaching<-gsub("[^0-9.-]", "", final_data$ti_teaching)
final_data$ti_administration<-gsub("[^0-9.-]", "", final_data$ti_administration)
final_data$ti_service_outreach<-gsub("[^0-9.-]", "", final_data$ti_service_outreach)


final_data$proposal_research<-gsub("[^0-9.-]", "", final_data$proposal_research)
final_data$proposal_teaching<-gsub("[^0-9.-]", "", final_data$proposal_teaching)
final_data$proposal_administration<-gsub("[^0-9.-]", "", final_data$proposal_administration)
final_data$proposal_service_outreach<-gsub("[^0-9.-]", "", final_data$proposal_service_outreach)

final_data$week_research<-gsub("[^0-9.-]", "", final_data$week_research)
final_data$week_teaching<-gsub("[^0-9.-]", "", final_data$week_teaching)
final_data$week_administration<-gsub("[^0-9.-]", "", final_data$week_administration)
final_data$week_service_outreach<-gsub("[^0-9.-]", "", final_data$week_service_outreach)

final_data$research<-gsub("[^0-9.-]", "", final_data$research)
final_data$research1<-gsub("[^0-9.-]", "", final_data$research1)
final_data$teaching<-gsub("[^0-9.-]", "", final_data$teaching)
final_data$teaching1<-gsub("[^0-9.-]", "", final_data$teaching1)
final_data$administration<-gsub("[^0-9.-]", "", final_data$administration)
final_data$administration1<-gsub("[^0-9.-]", "", final_data$administration1)
final_data$serviceoutreach<-gsub("[^0-9.-]", "", final_data$serviceoutreach)
final_data$service_outreach1<-gsub("[^0-9.-]", "", final_data$service_outreach1)
final_data$where_you_typically_wor<-toTitleCase(final_data$where_you_typically_wor)


final_data$disrupted_part[1]=5
final_data$disrupted_part[2]=NA
final_data$disrupted_part[3]=NA

final_data$my_physical_activity[2]=5
final_data$my_physical_activity[3]=5


# research_style1 = 'Hands-on [... plus experiments/analysis]'
# research_style2 = 'Hands-off [management/reading/writing]'
# 
# 
# final_data$test <-
#   case_when(
#     str_detect(final_data$test, research_style1) ~ "Hands-on",
#     str_detect(final_data$test, research_style2) ~
#       "Hands-off"
#   )

# colnames(final_data)

final_data<-final_data  %>%  dplyr::rename(record_id = "ï..record_id")

# final_data <- final_data[-c(89), ]
# # final_data <- final_data[-c(89), ]
final_data <- final_data[rownames(final_data) != "230",]
final_data <- final_data[rownames(final_data) != "288",]
final_data <- final_data[rownames(final_data) != "302",]
final_data <- final_data[rownames(final_data) != "649",]
final_data <- final_data[rownames(final_data) != "662",]


final_data$google_scholar_h_index[97]=18


# factor(final_data$what_is_the_average_number)

# final_data$email <- NULL
final_data$confirm_email <- NULL
final_data$standard_part_complete <- NULL
# colnames(final_data)

final_data$google_scholar_h_index[159] <- NA
final_data$google_scholar_h_index[238] <- NA
final_data$google_scholar_h_index[275] <- NA
final_data$google_scholar_h_index[334] <- NA
final_data$google_scholar_h_index[98] <- 20
final_data$google_scholar_h_index[244] <- NA
final_data$google_scholar_h_index[362] <- 63
final_data$google_scholar_h_index[210] <- NA
final_data$google_scholar_h_index[96] <- 198
final_data$google_scholar_h_index[142] <- NA
final_data$google_scholar_h_index[172] <- NA
final_data$google_scholar_h_index[307] <- NA
final_data$google_scholar_h_index[330] <- 119
final_data$google_scholar_h_index[120] <- 120


final_data$google_scholar_h_index[3] <- 15
final_data$google_scholar_h_index[109] <- 49
final_data$google_scholar_h_index[205] <- NA
final_data$google_scholar_h_index[225] <- 10
final_data$google_scholar_h_index[229] <- NA
final_data$google_scholar_h_index[236] <- 32



```

```{r, echo=FALSE}

final_data<- final_data[!is.na(final_data$google_scholar_h_index),]
final_data<- final_data[final_data$funding_proposal !="No",]

## only the data having comments on proposal 
# final_data <- filter(final_data, please_write_about_any_oth != "")
```

```{r, echo=FALSE}
# View(final_data)
final_data$email <- NULL
file_name='StandardPart.csv'
write.csv(final_data,file.path(curated_data_dir, file_name), row.names = FALSE)
```


```{r, echo=FALSE}
final_data$record_id <- NULL
for (i in 1:nrow(final_data)){
   if (final_data$wha_average_number[i]=="1-2"){
    final_data$wha_average_number[i]="1"
   }
  if (final_data$wha_average_number[i]=="3-4"){
    final_data$wha_average_number[i]="2"
   }
  if (final_data$wha_average_number[i]=="5-6"){
    final_data$wha_average_number[i]="3"
   }
  if (final_data$wha_average_number[i]=="7-9"){
    final_data$wha_average_number[i]="4"
  }
  if (final_data$wha_average_number[i]==">=10"){
    final_data$wha_average_number[i]="5"
  }
}

for (i in 1:nrow(final_data)){
   if (final_data$what_is_the_average_number[i]=="1-2"){
    final_data$what_is_the_average_number[i]="1"
   }
  if (final_data$what_is_the_average_number[i]=="3-4"){
    final_data$what_is_the_average_number[i]="2"
   }
  if (final_data$what_is_the_average_number[i]=="5-6"){
    final_data$what_is_the_average_number[i]="3"
   }
  if (final_data$what_is_the_average_number[i]=="7-9"){
    final_data$what_is_the_average_number[i]="4"
  }
  if (final_data$what_is_the_average_number[i]==">=10"){
    final_data$what_is_the_average_number[i]="5"
  }
}


final_data <- final_data  %>% dplyr::rename(
  Rank = "faculty_rank",
  Gender = "gender",
  Department = "departmental_affiliation",
  Department_Other = "please_write_down_the_depa",
  H = "google_scholar_h_index",
  State = "state_do_you_reside",
  Salary_Dependency_In_Month = "how_many_months_of_your_sa",
  Salary_Dependency_In_Month_Other = "please_specify_the_months",
  Research_Funded_By_External_Grants = "what_is_your_current_level",
  Deadline_Today = "do_you_have_a_deadline_tod",
  Workload_Today = "please_indicate_your_workl",
  Workplace = "where_you_typically_wor",
  Workplace_Other = "please_specify_your_work_p",
  RS = "test",  ## Research Style
  WH = "in_a_typical_workday_how_m",  ## Typical week working hours
  
  TWR = "ti_research",  ## Typical week Research
  TWT = "ti_teaching",  ## Typical week Research
  TWA = "ti_administration",  ## Typical week Research
  TWO = "ti_service_outreach",  ## Typical week Research
  
  TW_Sum = "sum_of_3",
  BR = "in_a_typical_day_how_often",  ## Break Frequency
  BR_Other = "please_specify_how_often_d",
  Email = "how_do_you_usually_handle",
  Proposal = "funding_proposal",
  NP = "what_is_the_average_number",  ## Average Number of Proposal per year
  Funding_Agency = "funding_agency",
  Funding_Agency_Other = "specify_the_name",
  P_SR = "hy",  ## Success Rate in the proposal
  AP = "how_far_in_advance_do_you",  # Composing Proposal
  P_AR = "what_is_the_length_of_supp",  ## Length of supporting research
  P_DWH = "in_the_week_leading_to_a_p",  ## Work load a week before Proposal Deadline
  
  P_DWR = "week_research",  ## Research percentage a week before Proposal
  P_DWT = "week_teaching",  ## Research percentage a week before Proposal
  P_DWA = "week_administration",  ## Research percentage a week before Proposal
  P_DWO = "week_service_outreach",  ## Research percentage a week before Proposal
  
  P_DW_SUm = "sum_of_4",
  
  PDR = "proposal_research",
  PDT = "proposal_teaching",
  PDA = "proposal_administration",
  PDO = "proposal_service_outreach",
  
  PD_Sum = "sum_of_5",

  
  
  
  # Research_DD="proposal_research",        ## Research percentage in the Proposal deadlines day
  P_T = "when_you_typically_subm",  ## Submit Proposal
  P_DS = "stress_level_in_a_fundi",  ## Stress leven on Proposal Deadline
  
  
  P_Disrupted_Research = "disrupted_part",
  P_Disrupted_Sleep = "my_sleep",
  P_Disrupted_Diet = "my_diet",
  P_Disrupted_PA = "my_physical_activity",
  P_Disrupted_IR = "my_interpersonal_relations",
  
  About_Proposal_Deadline = "please_write_about_any_oth",
  
  Conference = "refereed_conference",
  NCP = "wha_average_number",
  CR = "core_rank",
  C_SR = "if_you_submit_manuscripts",  
  ACP = "far_in_advance_do_you",  
  AR = "length_of_supp",  
  
  C_DWH = "in_the_week_leading_to_a_c",  
  
  C_DWR = "research",
  C_DWT = "teaching",
  C_DWA = "administration",
  C_DWO = "serviceoutreach",
  
  C_DW_Sum = "sum_of_6",
  
  CDR = "research1",
  CDT = "teaching1",
  CDA = "administration1",
  CDS = "service_outreach1",
  
  CD_Sum = "sum_of_7",
  
  DDR = "research1",
  C_T = "you_typically_subm",  ## Submit Proposal
  C_DS = "ss_level_in_a_fundi",     ## Stress leven on Proposal Deadline
  
  C_Disrupted_Research = "disrupted_conf",
  C_Disrupted_Sleep = "my_sleep2",
  C_Disrupted_Diet = "diet2",
  C_Disrupted_PA = "physical_activity2",
  C_Disrupted_IR = "interpersonal_relations2",
  
  About_Conf_Deadline = "please_write_about_any",
)
```

```{r, echo=FALSE}
# View(final_data)
# final_data$email <- NULL
file_name='AllData.csv'
write.csv(final_data,file.path(curated_data_dir, file_name), row.names = FALSE)
print(paste("The number of rows", nrow(final_data)))
```

```{r, echo=FALSE}
# levels(factor(final_data$Research_Funded_By_External_Grants))
# 
# 
# final_data <- final_data %>%
#   mutate(Pro_SR = recode(P_SR, "< 10%" = 0, "10-20%" = 0, "20-30%" = 0, "30-50%" = 1, "50-75%" = 1, "75-90%" = 1, "> 90%" = 1))
# table(final_data$Pro_SR)
```

```{r, echo=FALSE}
# final_data <- filter(final_data, P_SR == "1")
# file_name='Successful.csv'
# write.csv(final_data,file.path(curated_data_dir, file_name), row.names = FALSE)
```


```{r, echo=FALSE}
# final_data$P_ID <- paste0("P_", 1:403)
# final_data <- final_data %>%
#   mutate(
#     S_G30 = recode(
#       P_SR,
#       "< 10%" = 0,
#       "10-20%" = 0,
#       "20-30%" = 0,
#       "30-50%" = 1,
#       "50-75%" = 1,
#       "75-90%" = 1,
#       "> 90%" = 1
#     )
#   ) %>%
#   mutate(
#     S_G50 = recode(
#       P_SR,
#       "< 10%" = 0,
#       "10-20%" = 0,
#       "20-30%" = 0,
#       "30-50%" = 0,
#       "50-75%" = 1,
#       "75-90%" = 1,
#       "> 90%" = 1
#     )
#   ) %>%
#   mutate(
#     S_F75 = recode(
#       Research_Funded_By_External_Grants,
#       "Not funded" = 0,
#       "25-1%" = 0,
#       "50-25%" = 0,
#       "75-50%" = 0,
#       "100-75%" = 1,
#       "Fully funded" = 1
#     )
#   ) %>%
#   mutate(
#     S_FF = recode(
#       Research_Funded_By_External_Grants,
#       "Not funded" = 0,
#       "25-1%" = 0,
#       "50-25%" = 0,
#       "75-50%" = 0,
#       "100-75%" = 0,
#       "Fully funded" = 1
#     )
#   ) %>%
#   mutate(S_G50FF = case_when(S_G50 == 1 & S_FF == 1 ~ 1, TRUE ~ 0))
# 
# 
# Q_Q_DF <- subset(final_data, select=c("P_ID", "S_G30", "S_G50", "S_F75", "S_FF", "S_G50FF", "About_Proposal_Deadline"))
# Q_Q_DF$WC <- sapply(Q_Q_DF$About_Proposal_Deadline, function(x) length(unlist(strsplit(as.character(x), "\\W+"))))
# Q_Q_DF <- Q_Q_DF %>% mutate(Comment = case_when(WC != 0 ~ 1, TRUE ~ 0))
# 
# 
# 
# 
# 
# Q_Q_DF_comment_only  <-
#   filter(Q_Q_DF, About_Proposal_Deadline != "")
# 
# Q_Q_DF_comment_only <- subset(Q_Q_DF_comment_only, 
#     P_ID != 'P_100' &
#       P_ID != 'P_122' &
#       P_ID != 'P_123' &
#       P_ID != 'P_182' &
#       P_ID != 'P_210' & 
#       P_ID != 'P_377' & 
#       P_ID != 'P_384'
#   )
# 
# 
# # library("xlsx")
# data_file_name <- 'Reviewed_file.xlsm'
# # Q_Q_freefomat<-read.xlsx(file.path(curated_data_dir,data_file_name),stringsAsFactors = FALSE)
# Q_Q_freefomat<-read.xlsx(file.path(curated_data_dir,data_file_name))
# 
# QQ_Bridge <- merge(Q_Q_DF_comment_only, Q_Q_freefomat, by = c("P_ID", "S_G30", "S_G50", "S_F75", "S_FF", "S_G50FF"), all.x = TRUE)
# 
# QQ_Bridge$About_Proposal_Deadline.x<- NULL
# QQ_Bridge$About_Proposal_Deadline.y<- NULL
# 
# 
# 
# 
# file_name='QQ_Bridge.csv'
# write.csv(QQ_Bridge,file.path(curated_data_dir, file_name), row.names = FALSE)
```
