==> scripts/3-dc-corequestions-processing.rmd <==
---
title: "Core Questions Processing"
output:
  html_document:
    df_print: paged
  pdf_document: default
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(dplyr)
library(freqtables)
library(plotrix)
library(plotly)
library(tools)
library(plotrix)
library(plotly)
library(grid)
library(gridExtra)
library(ggpubr)
library(cowplot)
library(reshape2)
library(plyr)
library(ggplot2)
library(dplyr)
library(plotrix)
library(plotly)
library(grid)
library(gridExtra)
library(ggpubr)
library(cowplot)
library(reshape2)
library(plyr)
library(ggplot2)
library(ggrepel)
library(forcats)
library(scales)
library(stringr)
# https://cran.r-project.org/web/packages/freqtables/vignettes/descriptive_analysis.html
# Core_Questions %>% freq_table(R_Style)
```

```{r, echo=FALSE}
root_dir <- getwd()
project_directory <- dirname(root_dir)
data_dir <- file.path(project_directory, 'curated-data')
plot_dir <- file.path(project_directory, 'Plots')
data_file_name <- 'StandardPart.csv'
curated_data_dir <- file.path(project_directory, 'curated-data')
StandardPart <-
  read.csv(file.path(data_dir, data_file_name), stringsAsFactors = FALSE)

```

```{r, echo=FALSE}

for (i in 1:nrow(StandardPart)) {
  if (StandardPart$departmental_affiliation[i] == "Other") {
    StandardPart$departmental_affiliation[i] = StandardPart$please_write_down_the_depa[i]
  }

}
```



```{r, echo=FALSE}
title_list = c(
  "Faculty Rank",
  "Gender",
  "Departments",
  "States",
  "Salary dependance on external grants",
  "Percentage of research operations funded by external grants",
  "Any looming deadline- Today/Next couple of days?",
  "Today's workload",
  "Workplace",
  "Research style",
  "Working hours- In a typical week",
  "Break frequency",
  "Email response style",
  "Participants submitting funding proposals",
  "Average number of proposals per year",
  "Funding agencies",
  "Perceived proposal success rate",
  "Length of proposal composition period",
  "Length of supporting research for proposals",
  "Proposal deadline week workload wrt a typical week",
  "Timing of proposal submission",
  "Stress level on proposal deadlines wrt regular working days",
  "Participants submitting papers in refereed conferences",
  "Average number of conference papers per year",
  "CORE rank of conferences",
  "Perceived success rate in  conference submissions",
  "Length of conference paper composition period",
  "Length of supporting research for conference papers",
  "Conference deadline week workload wrt a typical week",
  "Timing of conference paper submission",
  "Stress level on conference deadlines wrt regular working days"
)
```





```{r, echo=FALSE, warning=FALSE}
Core_Questions <-
  StandardPart %>% dplyr::select(
    faculty_rank,
    gender,
    departmental_affiliation,
    record_id,
    state_do_you_reside,
    how_many_months_of_your_sa,
    what_is_your_current_level,
    do_you_have_a_deadline_tod,
    please_indicate_your_workl,
    where_you_typically_wor,
    test,
    in_a_typical_workday_how_m,
    ti_research,
    in_a_typical_day_how_often,
    how_do_you_usually_handle,
    funding_proposal,
    what_is_the_average_number,
    funding_agency,
    hy,
    how_far_in_advance_do_you,
    what_is_the_length_of_supp,
    in_the_week_leading_to_a_p,
    when_you_typically_subm,
    stress_level_in_a_fundi,
    refereed_conference,
    wha_average_number,
    core_rank,
    if_you_submit_manuscripts,
    far_in_advance_do_you,
    length_of_supp,
    in_the_week_leading_to_a_c,
    you_typically_subm,
    ss_level_in_a_fundi,
    please_write_about_any_oth
  )



Core_Questions <-
  Core_Questions  %>%  dplyr::rename(
    FR = "faculty_rank",
    Gender = "gender",
    Departmnet = "departmental_affiliation",
    FA = "funding_agency",
    Record_id = "record_id",
    State = "state_do_you_reside",
    S_D_E = "how_many_months_of_your_sa",
    ## salary depends on external grants
    RO_F_EG = "what_is_your_current_level",
    ## Research operations funded by external grants
    Deadline_today = "do_you_have_a_deadline_tod",
    Workload_today = "please_indicate_your_workl",
    Workplace = "where_you_typically_wor",
    RS = "test",
    ## Research Style
    TW_W_H = "in_a_typical_workday_how_m",
    ## Typical week working hours
    
    TW_Research="ti_research",              ## Typical week Research
    Break = "in_a_typical_day_how_often",
    ## Break Frequency
    Email = "how_do_you_usually_handle",
    ## Email respnse style
    NP = "what_is_the_average_number",
    ## Average Number of Proposal per year
    Success = "hy",
    ## Success Rate in the proposal
    Com_Proposal = "how_far_in_advance_do_you",
    # Composing Proposal
    L_Of_SR = "what_is_the_length_of_supp",
    ## Length of supporting research
    W_WB_PD = "in_the_week_leading_to_a_p",
    ## Work load a week before Proposal Deadline
    # Research_WB="week_research",            ## Research percentage a week before Proposal
    # Research_DD="proposal_research",        ## Research percentage in the Proposal deadlines day
    T = "when_you_typically_subm",
    ## Submit Proposal
    DS = "stress_level_in_a_fundi",
    ## Stress leven on Proposal Deadline
    A_N_Conf_Pap = "wha_average_number",
    Comments = "please_write_about_any_oth"
  )


# levels(factor(Core_Questions$State))

East = 'Connecticut|Maryland|Massachusetts|^New Jersey$|^New York$|Pennsylvania|^Virginia$|^North Carolina$'
Midwest = 'Illinois|Michigan|Minnesota|Ohio|Wisconsin'
West = 'Arizona|California|Colorado|Nevada|^New Mexico$|Oregon|^Utah$'
South = 'Florida|Georgia|Tennessee|Texas|Alabama|^Oklahoma$'

Core_Questions <- Core_Questions %>%
  mutate(
    States_ewns = case_when(
      str_detect(State, East) ~ 'East',
      str_detect(State, West) ~ 'West',
      str_detect(State, Midwest) ~ 'Midwest',
      str_detect(State, South) ~ 'South'
    )
  )
Core_Questions$State <- Core_Questions$States_ewns

Core_Questions$States_ewns <- NULL

```



```{r, echo=FALSE}
# count(Core_Questions$Break)
for(i in 1:nrow(Core_Questions)){
  
  #######---------Salary depends of external grants------------#################
  if (Core_Questions$S_D_E[i]=="Three months"){
    Core_Questions$S_D_E[i]="Three months"
  }else if (Core_Questions$S_D_E[i]=="2 months" || Core_Questions$S_D_E[i]==2){
    Core_Questions$S_D_E[i]="Two months"
  }else{
    Core_Questions$S_D_E[i]="Other"
  }


  if (Core_Questions$RS[i] == "Hands-off") {
    Core_Questions$RS[i] = "RS1"
  } else if (Core_Questions$RS[i] == "") {
    Core_Questions$RS[i] = NA
  } else{
    Core_Questions$RS[i] = "RS2"
  }
  
  
  if (Core_Questions$NP[i] == "1-2") {
    Core_Questions$NP[i] = "NP1"
  } else if (Core_Questions$NP[i] == "3-4") {
    Core_Questions$NP[i] = "NP2"
  } else if (Core_Questions$NP[i] == "5-6") {
    Core_Questions$NP[i] = "NP3"
  } else if (Core_Questions$NP[i] == "7-9") {
    Core_Questions$NP[i] = "NP3"
  } else if (Core_Questions$NP[i] == "") {
    Core_Questions$NP[i] = NA
  } else {
    Core_Questions$NP[i] = "NP3"
  }
  
  
  if (Core_Questions$RO_F_EG[i]=="Not funded"){
    Core_Questions$RO_F_EG[i]="Not funded"
  } else if (Core_Questions$RO_F_EG[i]=="25-1%"){
    Core_Questions$RO_F_EG[i]="1-25%"
  } else if (Core_Questions$RO_F_EG[i]=="50-25%"){
    Core_Questions$RO_F_EG[i]="25-50%"
  } else if (Core_Questions$RO_F_EG[i]=="75-50%"){
    Core_Questions$RO_F_EG[i]="50-75%"
  } else if (Core_Questions$RO_F_EG[i]=="100-75%"){
    Core_Questions$RO_F_EG[i]="75-100%"
  } else{
    Core_Questions$RO_F_EG[i]="Fully funded"
  }
  
  
  if (Core_Questions$Workplace[i]=="Home"){
    Core_Questions$Workplace[i]="Home"
  } else if (Core_Questions$Workplace[i]=="Office"){
    Core_Questions$Workplace[i]="Office"
  } else {
    Core_Questions$Workplace[i]="Other"
  }

  if (Core_Questions$Email[i]=="Answer email in batches once  or twice per day"){
    Core_Questions$Email[i]="Reply Once/Twice"
  }else {
    Core_Questions$Email[i]="Reply instantly"
  }

  
  
  
  if (Core_Questions$Break[i]=="Every 1-2 hours"){
    Core_Questions$Break[i]="Every 1-2 hours"
  } else if (Core_Questions$Break[i]=="Every 3-4 hours"){
    Core_Questions$Break[i]="Every 3-4 hours"
  } else if (Core_Questions$Break[i]=="Every hour"){
    Core_Questions$Break[i]="Every hour"
  } else{
    Core_Questions$Break[i]="Other"
  }
  
  
  
  if (Core_Questions$FA[i]=="NSF"){
    Core_Questions$FA[i]="NSF"
  } else if (Core_Questions$FA[i]=="NIH"){
    Core_Questions$FA[i]="NIH"
  } else if (Core_Questions$FA[i]=="DOE"){
    Core_Questions$FA[i]="DOE"
  } else if (Core_Questions$FA[i]=="DOD"){
    Core_Questions$FA[i]="DOD"
  } else if (Core_Questions$FA[i]=="NASA"){
    Core_Questions$FA[i]="NASA"
  }  else{
    Core_Questions$FA[i]="Other"
  }
  
  
  # if (Core_Questions$T[i]=="Minutes before deadline"){
  #   Core_Questions$T[i]="Minutes before"
  # } else if (Core_Questions$T[i]=="1-3 hours before deadline"){
  #   Core_Questions$T[i]="1-3 hours before"
  # } else if (Core_Questions$T[i]=="3-6 hours before deadline"){
  #   Core_Questions$T[i]="3-6 hours before"
  # }  else if (Core_Questions$T[i]=="1 day before deadline"){
  #   Core_Questions$T[i]="1 day before"
  # } else if (Core_Questions$T[i]==""){
  #   Core_Questions$T[i]=NA
  # } else {
  #   Core_Questions$T[i]="2 or more days before"
  # }
  
  
  if (Core_Questions$T[i] == "Minutes before deadline") {
    Core_Questions$T[i] = "T1"
  } else if (Core_Questions$T[i] == "1-3 hours before deadline") {
    Core_Questions$T[i] = "T1"
  } else if (Core_Questions$T[i] == "3-6 hours before deadline") {
    Core_Questions$T[i] = "T1"
  } else if (Core_Questions$T[i] == "1 day before deadline") {
    Core_Questions$T[i] = "T2"
  } else if (Core_Questions$T[i] == "") {
    Core_Questions$T[i] = NA
  } else {
    Core_Questions$T[i] = "T2"
  }
  
    if (Core_Questions$you_typically_subm[i]=="Minutes before deadline"){
    Core_Questions$you_typically_subm[i]="Minutes before"
  } else if (Core_Questions$you_typically_subm[i]=="1-3 hours before deadline"){
    Core_Questions$you_typically_subm[i]="1-3 hours before"
  } else if (Core_Questions$you_typically_subm[i]=="3-6 hours before deadline"){
    Core_Questions$you_typically_subm[i]="3-6 hours before"
  }  else if (Core_Questions$you_typically_subm[i]=="1 day before deadline"){
    Core_Questions$you_typically_subm[i]="1 day before"
  } else if (Core_Questions$you_typically_subm[i]==""){
    Core_Questions$you_typically_subm[i]=NA
  } else {
    Core_Questions$you_typically_subm[i]="2 or more days before"
  }
  
  
  if (Core_Questions$DS[i] == "Extremely less"){
    Core_Questions$DS[i] = "Significantly less"
  } else if (Core_Questions$DS[i] == "Significantly less"){
    Core_Questions$DS[i] = "Less"
  } else if (Core_Questions$DS[i] == "Same"){
    Core_Questions$DS[i] = "About the same"
  } else if (Core_Questions$DS[i] == "Significantly  more"){
    Core_Questions$DS[i] = "More"
  } else if (Core_Questions$DS[i] == "Extremely  more"){
    Core_Questions$DS[i] = "Significantly more"
  } else {
    Core_Questions$DS[i] = NA
  }
  

    if (Core_Questions$ss_level_in_a_fundi[i] == "Extremely less"){
    Core_Questions$ss_level_in_a_fundi[i] = "Significantly less"
  } else if (Core_Questions$ss_level_in_a_fundi[i] == "Significantly less"){
    Core_Questions$ss_level_in_a_fundi[i] = "Less"
  } else if (Core_Questions$ss_level_in_a_fundi[i] == "Same"){
    Core_Questions$ss_level_in_a_fundi[i] = "About the same"
  } else if (Core_Questions$ss_level_in_a_fundi[i] == "Significantly  more"){
    Core_Questions$ss_level_in_a_fundi[i] = "More"
  } else if (Core_Questions$ss_level_in_a_fundi[i] == "Extremely  more"){
    Core_Questions$ss_level_in_a_fundi[i] = "Significantly more"
  } else {
    Core_Questions$ss_level_in_a_fundi[i] = NA
  }
  
  
}


file_name = "CoreQuestions.csv"
write.csv(Core_Questions,file.path(curated_data_dir, file_name), row.names = FALSE)
```


<!--```{r, echo=FALSE}

list1 = c("East", "West", "Midwest", "South")
list2 = c("Two months", "Three months", "Other")

list3 = c("Not funded",
          "1-25%",
          "25-50%",
          "50-75%",
          "75-100%",
          "Fully funded")
list4 = c("No", "Yes")

list5 = c("Light", "Standard", "Heavy")
list6 = c("Home", "Office", "Other")

# list7 = c("Hands-off", "Hands-on")
list7 = c("RS1", "RS2")
list8 = c("< 30", "30-40", "40-50", "> 50")

list9 = c("Every hour", "Every 1-2 hours", "Every 3-4 hours", "Other")
list10 = c("Reply instantly", "Reply Once/Twice")

list11 = c("Yes")
list12 = c("1-2", "3-4", "5-6", "7-9", ">=10")
# list12 = c("NP1", "NP2", "NP3")

list13 = c("NSF", "NIH", "DOE","DOD", "NASA",  "Other")
list14 = c("< 10%", "10-20%", "20-30%", "30-50%", "50-75%", "75-90%", "> 90%")

list15 = c("1-2 weeks", "2-4 weeks", "1-2 months", "> 2 months")
list16 = c("< 1 month", "1-3 months", "3-6 months", "6-12 months", "> 12 months")

list17 = c("Significantly less", "Less", "About the same", "More", "Significantly more")
# list18 = c("Minutes before",  "1-3 hours before",  "3-6 hours before",  "1 day before",  "2 or more days before")
list18 = c("T1", "T2")

list19 = c("Significantly less", "Less", "About the same", "More", "Significantly more")
# list19 = c("DS1", "DS2")

list20 = c("No", "Yes")

list21 = c("1-2", "3-4", "5-6", "7-9", ">= 10")
list22 = c("C", "B", "A", "A*")

list23 = c("< 10%", "10-20%", "20-30%", "30-50%", "50-75%", "75-90%", "> 90%")
list24 = c("< 1 week", "1-2 weeks", "2-4 weeks", "1-2 months", "> 2 months")

list25 = c("< 1 month", "1-3 months", "3-6 months", "6-12 months", "> 12 months")
list26 = c( "Significantly less","Less","About the same", "More", "Significantly more")

list27 = c(
  "Minutes before",
  "1-3 hours before",
  "3-6 hours before",
  "1 day before",
  "2 or more days before"
)

list28 = c("Significantly less", "Less", "About the same", "More", "Significantly more")

Ticks_list <-
  list(list1,
       list2,
       list3,
       list4,
       list5,
       list6,
       list7,
       list8,
       list9,
       list10,
       list11,
       list12,
       list13,
       list14,
       list15,
       list16,
       list17,
       list18,
       list19,
       list20,
       list21,
       list22,
       list23,
       list24,
       list25,
       list26,
       list27,
       list28)


plot_list <- list()

for (i in 2:length(Core_Questions)) {
# for (i in 2:3) {
  # for (i in 18:19) {
  ylimit = 250
  if ((i == 3) |
      (i == 5) | (i == 8) | (i == 11) | (i == 12) | (i == 14) | (i == 20)) {
    ylimit = 400
  }
  temp <- count(Core_Questions[i])
  colnames(temp) <- c("item", "count")
  temp <- temp[!(temp$item == ""), ]
  temp <- temp[order(temp$count),]
  temp <- temp[complete.cases(temp), ]
 
    
  temp2 = Core_Questions %>% freq_table(Core_Questions[i])
  temp2 = temp2[, c(2, 8, 9)]
  temp2 <- temp2[!(temp2$cat == ""), ]
  temp2 <- temp2[complete.cases(temp2), ]
  temp2<-temp2  %>%  dplyr::rename( Item="cat", LCI="lcl", UCI="ucl") 
  temp2$LCI <- format(round(temp2$LCI, 2), nsmall = 2)
  temp2$UCI <- format(round(temp2$UCI, 2), nsmall = 2)
  
  x_lebel = paste("\n\n 95% CI ")
  
  for (row_num in 1:nrow(temp2)){
    if (row_num == 3){
      space = "\n"
    } else {
      space = "      "
    }
    x_lebel = paste0(x_lebel, temp2$Item[row_num], "(",temp2$LCI[row_num], ", " ,temp2$UCI[row_num], ")", space)
  }
  
  
  bar_plot <- ggplot(data = temp, aes(x = item, y = count)) +
    geom_bar(stat = "identity",
             width = 0.5,
             fill = "steelblue") +
    geom_text(aes(label=count), vjust=1, color="white", size=3.5) +
    theme_bw() +
    scale_y_continuous(breaks = seq(0, ylimit, by = 50),
                       limits = c(0, ylimit)) +
    labs(x = x_lebel, y = "Count", title = title_list[i - 1]) +
    theme(
      text = element_text(size = 14),
      panel.grid.major = element_blank(),
      panel.grid.minor = element_blank(),
      plot.title = element_text(hjust = 0.5),
      axis.text.x = element_text(
        face = "bold",
        size = 14 ,
        angle = 30,
        hjust = 1
      ),
      axis.text.y = element_text(face = "bold", size = 14)
    ) + scale_x_discrete(limits = Ticks_list[[i - 1]])
  
  
  
  plot_list[[length(plot_list) + 1]] <- bar_plot
  

  
  temp <- temp %>%
    mutate(
      cs = rev(cumsum(rev(count))),
      prop = percent(count / sum(count)),
      pos = count / 2 + lead(cs, 1),
      pos = if_else(is.na(pos), count / 2, pos)
    )
  temp$pos[is.na(temp$pos)] <- 1
  
  
  pichart <-
    ggplot(temp, aes(
      x = "" ,
      y = count,
      fill = fct_inorder(item)
    )) +
    geom_col(width = 1) +
    coord_polar(theta = "y", start = 0) +
    scale_fill_brewer(palette = "Set3", direction = -4) +
    geom_label_repel(
      aes(y = pos, label = paste0(item, " " , prop)),
      data = temp,
      size = 4,
      show.legend = F,
      nudge_x = 1
    ) +
    theme_void() +
    theme(legend.position = "none", legend.title = element_blank())
  
  plot_list[[length(plot_list) + 1]] <- pichart
  
}
as.character("state_do_you_reside")

plots_act = ggarrange(plotlist = plot_list,
                      nrow = 2,
                      ncol = 1)

for (i in 1:length(plots_act)) {
  final_plot <- plot_grid(plots_act[[i]])
  filename <- paste("plot", i, ".pdf", sep = "")
  full_path <- file.path(plot_dir, filename)
  ggsave(
    full_path,
    final_plot,
    width = 8.5,
    height = 11,
    units = "in"
  )
  print(final_plot)
}



  # temp2 = Core_Questions %>% freq_table(Core_Questions$R_Style)
  # temp2 = temp2[, c(2, 8, 9)]
  # temp2<-temp2  %>%  dplyr::rename( Item="cat", LCI="lcl", UCI="ucl")
Data_CQ <- Core_Questions[,c(13, 16, 9, 10, 18, 20)]

file_name='Selected_Core_Questions.csv'
write.csv(Data_CQ,file.path(curated_data_dir, file_name), row.names = FALSE)

library(ggplot2)
Funding_WRT_NP <- ggplot(Core_Questions, aes(x = RO_F_EG)) + geom_histogram(stat = "count") +
  facet_grid( ~ NP) +
  theme(
    axis.title.y = element_text(size=12, face="bold", colour = "black"),
    panel.grid.major = element_blank(),
    panel.grid.minor = element_blank(),
    plot.title = element_text(hjust = 0.5),
    axis.text.x = element_text(
      face = "bold",
      size = 10 ,
      angle = 30,
      hjust = 1
    ),
    axis.text.y = element_text(
      face = "bold",
      size = 10), 
    strip.text = element_text(size = 12, face = "bold")
  ) +
  labs(x = "", y = "Count")


filename <- "Funding_WRT_NP.pdf"
full_path <- file.path(plot_dir, filename)
ggsave(
  full_path,
  Funding_WRT_NP,
  width = 8.5,
  height = 5.5,
  units = "in"
)

Funding_NP <-Core_Questions[, c("RO_F_EG", "NP")]
NP<- c("NP1", "NP2", "NP3")

plot_list<-list()
for (item in NP){
  print(item)
  sub_np <- subset(Funding_NP, NP==item)

  temp <- count(sub_np)
  temp$NP <- NULL
  colnames(temp) <- c("item", "count")
  temp <- temp %>%
    mutate(
      cs = rev(cumsum(rev(count))),
      prop = percent(count / sum(count)),
      pos = count / 2 + lead(cs, 1),
      pos = if_else(is.na(pos), count / 2, pos)
    )
  temp$pos[is.na(temp$pos)] <- 1


  pichart <-
    ggplot(temp, aes(
      x = "" ,
      y = count,
      fill = fct_inorder(item)
    )) +
    geom_col(width = 1) +
    coord_polar(theta = "y", start = 0) +
    scale_fill_brewer(palette = "Set3", direction = -4) +
    geom_label_repel(
      aes(y = pos, label = paste0(item, " " , prop)),
      data = temp,
      size = 4,
      show.legend = F,
      nudge_x = 1
    ) +
    theme_void() +
    theme(legend.position = "none", legend.title = element_blank())
  
  plot_list[[length(plot_list) + 1]] <- pichart
  print(pichart)
}

pie_chart=ggarrange(plotlist = plot_list,
                      nrow = 1,
                      ncol = 3)


final_plot <- ggarrange(Funding_WRT_NP, pie_chart, nrow = 2, ncol = 1)

filename <- "Funding_WRT_NP_both.pdf"
full_path <- file.path(plot_dir, filename)
ggsave(
  full_path,
  final_plot,
  width = 8.5,
  height = 5.5,
  units = "in"
)
  



library(ggplot2)
Funding_FA_NP <- ggplot(Core_Questions, aes(x = FA)) + geom_histogram(stat = "count") +
  facet_grid( ~ NP) +
  theme(
    axis.title.y = element_text(size=12, face="bold", colour = "black"),
    panel.grid.major = element_blank(),
    panel.grid.minor = element_blank(),
    plot.title = element_text(hjust = 0.5),
    axis.text.x = element_text(
      face = "bold",
      size = 10 ,
      angle = 30,
      hjust = 1
    ),
    axis.text.y = element_text(
      face = "bold",
      size = 10), 
    strip.text = element_text(size = 12, face = "bold")
  ) +
  labs(x = "", y = "Count")


Funding_FA_NP

FA_NP <-Core_Questions[, c("FA", "NP")]
NP<- c("NP1", "NP2", "NP3")

plot_list<-list()
for (item in NP){
  print(item)
  sub_np <- subset(FA_NP, NP==item)
  temp <- count(sub_np)
  temp$NP <- NULL
  colnames(temp) <- c("item", "count")
  
  temp <- temp %>%
    mutate(
      cs = rev(cumsum(rev(count))),
      prop = percent(count / sum(count)),
      pos = count / 2 + lead(cs, 1),
      pos = if_else(is.na(pos), count / 2, pos)
    )
  temp$pos[is.na(temp$pos)] <- 1

  # temp <- temp[order(temp$pos),]

  pichart <-
    ggplot(temp, aes(
      x = "" ,
      y = count,
      fill = fct_inorder(item)
    )) +
    geom_col(width = 1) +
    coord_polar(theta = "y", start = 0) +
    scale_fill_brewer(palette = "Set3", direction = -4) +
    geom_label_repel(
      aes(y = pos, label = paste0(item, " " , prop)),
      data = temp,
      size = 4,
      show.legend = F,
      nudge_x = 0.5
    ) +
    theme_void() +
    theme(legend.position = "none", legend.title = element_blank())
  
  plot_list[[length(plot_list) + 1]] <- pichart
  print(pichart)
}

pie_chart=ggarrange(plotlist = plot_list,
                      nrow = 1,
                      ncol = 3)


final_plot <- ggarrange(Funding_FA_NP, pie_chart, nrow = 2, ncol = 1)

filename <- "Funding_Agency_NP_both.pdf"
# filename <- "Funding_Agency_NP_both.jpg"
full_path <- file.path(plot_dir, filename)
ggsave(
  full_path,
  final_plot,
  width = 8.5,
  height = 5.5,
  units = "in"
)
  
library(ggplot2)
RS_NP <- ggplot(Core_Questions, aes(x = RS)) + geom_histogram(stat = "count") +
  facet_grid( ~ NP) +
  theme(
    axis.title.y = element_text(size=12, face="bold", colour = "black"),
    panel.grid.major = element_blank(),
    panel.grid.minor = element_blank(),
    plot.title = element_text(hjust = 0.5),
    axis.text.x = element_text(
      face = "bold",
      size = 10 ,
      angle = 30,
      hjust = 1
    ),
    axis.text.y = element_text(
      face = "bold",
      size = 10), 
    strip.text = element_text(size = 12, face = "bold")
  ) +
  labs(x = "", y = "Count")


RS_NP

FA_NP <-Core_Questions[, c("RS", "NP")]
NP<- c("NP1", "NP2", "NP3")

plot_list<-list()
for (item in NP){
  print(item)
  sub_np <- subset(FA_NP, NP==item)
  temp <- count(sub_np)
  temp$NP <- NULL
  colnames(temp) <- c("item", "count")
  
  temp <- temp %>%
    mutate(
      cs = rev(cumsum(rev(count))),
      prop = percent(count / sum(count)),
      pos = count / 2 + lead(cs, 1),
      pos = if_else(is.na(pos), count / 2, pos)
    )
  temp$pos[is.na(temp$pos)] <- 1

  # temp <- temp[order(temp$pos),]

  pichart <-
    ggplot(temp, aes(
      x = "" ,
      y = count,
      fill = fct_inorder(item)
    )) +
    geom_col(width = 1) +
    coord_polar(theta = "y", start = 0) +
    scale_fill_brewer(palette = "Set3", direction = -4) +
    geom_label_repel(
      aes(y = pos, label = paste0(item, " " , prop)),
      data = temp,
      size = 4,
      show.legend = F,
      nudge_x = 0.5
    ) +
    theme_void() +
    theme(legend.position = "none", legend.title = element_blank())
  
  plot_list[[length(plot_list) + 1]] <- pichart
  print(pichart)
}

pie_chart=ggarrange(plotlist = plot_list,
                      nrow = 1,
                      ncol = 3)


final_plot <- ggarrange(RS_NP, pie_chart, nrow = 2, ncol = 1)

filename <- "RS_NP_both.pdf"
# filename <- "Funding_Agency_NP_both.jpg"
full_path <- file.path(plot_dir, filename)
ggsave(
  full_path,
  final_plot,
  width = 8.5,
  height = 5.5,
  units = "in"
)
```-->
