==> scripts/4-dc-keydata-processing.rmd <==
---
title: "Key Data processing"
output:
  html_document:
    df_print: paged
  pdf_document: default
---

```{r setup, include=FALSE}
library(dplyr)
library(ggplot2)
library(GGally)
library(reshape2)
library(plyr)
library(stringr)
knitr::opts_chunk$set(echo = TRUE)
```


```{r, echo=FALSE}
root_dir <- getwd()
project_directory <- dirname(root_dir)
data_dir <- file.path(project_directory, 'raw-data')
curated_data_dir <- file.path(project_directory, 'curated-data')
plot_dir <- file.path(project_directory, 'Plots')
data_file_name <- 'Procesed_Psychometric_Data.csv'
data_file_name_std <- 'StandardPart.csv'


Data <-
  read.csv(file.path(curated_data_dir, data_file_name),
           stringsAsFactors = FALSE)
Raw_Data <-
  read.csv(file.path(curated_data_dir, data_file_name_std),
           stringsAsFactors = FALSE)
colnames(Data) <-
  c(
    "Rank",
    "Mental",
    "Physical",
    "Temporal",
    "Performance",
    "Effort",
    "Frustration",
    "NASA",
    "TA",
    "EXT",
    "AGR",
    "CS",
    "NT",
    "OP",
    "AV",
    "EM",
    "Task"
  )
```


```{r, echo=FALSE}
Corelation_DF <-
  data.frame(Data[, c(1, 8:17)], Raw_Data[, c(4:7, 15:17, 22, 26:27, 29:33, 43, 44, 10)])




for (i in 1:nrow(Corelation_DF)) {
  if (Corelation_DF$departmental_affiliation[i] == "Other") {
    Corelation_DF$departmental_affiliation[i] = Corelation_DF$please_write_down_the_depa[i]
  }

}

Corelation_DF$please_write_down_the_depa <- NULL


Corelation_DF <-
  Corelation_DF  %>%  dplyr::rename(
    H = "google_scholar_h_index",
    RS = "test",
    Departmnet = departmental_affiliation,
    State = state_do_you_reside,
    ## Research Style
    WH = "in_a_typical_workday_how_m",
    ## Typical week working hours
    TWR = "ti_research",
    FA = "funding_agency",
    ## Typical week Research
    BF = "in_a_typical_day_how_often",
    ## Break Frequency
    NP = "what_is_the_average_number",
    ## Average Number of Proposal per year
    SR = "hy",
    ## Success Rate in the proposal
    AP = "how_far_in_advance_do_you",
    # Composing Proposal
    AR = "what_is_the_length_of_supp",
    ## Length of supporting research
    DWH = "in_the_week_leading_to_a_p",
    ## Work load a week before Proposal Deadline
    DWR = "week_research",
    ## Research percentage a week before Proposal
    # Research_DD="proposal_research",        ## Research percentage in the Proposal deadlines day
    T = "when_you_typically_subm",
    ## Submit Proposal
    DS = "stress_level_in_a_fundi",     ## Stress leven on Proposal Deadline
    FC= "what_is_your_current_level"
  )


Corelation_DF <- Corelation_DF %>%
  mutate(FC = recode(FC, "Not funded" = 1, "25-1%" = 2, "50-25%" = 3, "75-50%" = 4, "100-75%" = 5, "Fully funded" = 6))

```


```{r, echo=FALSE}
Natural_Sciences = '^Applied Math$|^Chemistry and Biochemistry$|^Oceanography$|^Environmental Studies$|^Chemistry$|^Physics$|^Geosciences$|^Mathematics$'
CIS = '^Informatics$|^Information$|^Information and Logistics Technology$|^Information Science$|^Computer Science$|^Scientific Computing$'
Engineering = '^Engineering / Learning Sciences$|^Management$|^Statistical Sciences and Operations Research$|^Technology$|^Engineering$'
Biomedical_Sciences = '^Health$|^Health and Human Performance$|^HHP$|^Neurobiology$|^Neuroscience$|^Optometry$|^Pharmacological and Pharmaceutical Sciences$|^Pharmacy$|^Vision Science$|^Biology$|^Medicine$|^Pharmacy $'
Behavioral_Sciences = '^Political Science$|^Sociology$|^Psychology$|^Speech, Language, and Hearing$|^Political Science $|^Communication Studies$'





Corelation_DF <- Corelation_DF %>%
  mutate(
    Disciplines = case_when(
      # str_detect(Departmnet, Natural_Sciences) ~ 'Natural Sciences',
      str_detect(Departmnet, Natural_Sciences) ~ 'NAT',
      # str_detect(Departmnet, CIS) ~ 'Computer & Information Sciences',
      str_detect(Departmnet, CIS) ~ 'CIS',
      # str_detect(Departmnet, Engineering) ~ 'Engineering',
      str_detect(Departmnet, Engineering) ~ 'ENG',
      # str_detect(Departmnet, Biomedical_Sciences) ~ 'Biomedical Sciences',
      str_detect(Departmnet, Biomedical_Sciences) ~ 'BIO',
      # str_detect(Departmnet, Behavioral_Sciences) ~ 'Behavioral Sciences'
      str_detect(Departmnet, Behavioral_Sciences) ~ 'BEHAV'
    )
  )
```





```{r, echo=FALSE}

East = 'Connecticut|Maryland|Massachusetts|^New Jersey$|^New York$|Pennsylvania|^Virginia$|^North Carolina$'
Midwest = 'Illinois|Michigan|Minnesota|Ohio|Wisconsin'
West = 'Arizona|California|Colorado|Nevada|^New Mexico$|Oregon|^Utah$'
South = 'Florida|Georgia|Tennessee|Texas|Alabama|^Oklahoma$'

Corelation_DF <- Corelation_DF %>%
  mutate(
    State_Geo_location = case_when(
      str_detect(State, East) ~ 'East',
      str_detect(State, West) ~ 'West',
      str_detect(State, Midwest) ~ 'Midwest',
      str_detect(State, South) ~ 'South'
    )
  )
Corelation_DF$State <- Corelation_DF$State_Geo_location


```


```{r, echo=FALSE}
# count(Corelation_DF$Stress_PD)
for (i in 1:nrow(Corelation_DF)) {
  #####-------Research Style-------------####################
  if (Corelation_DF$RS[i] == "Hands-off") {
    Corelation_DF$RS[i] = 1
  } else{
    Corelation_DF$RS[i] = 2
  }
  
  ##########-----Typical Week working Hours-------------#############
  if (Corelation_DF$WH[i] == "< 30") {
    Corelation_DF$WH[i] = 1
  } else if (Corelation_DF$WH[i] == "30-40") {
    Corelation_DF$WH[i] = 1
  } else if (Corelation_DF$WH[i] == "40-50") {
    Corelation_DF$WH[i] = 2
  } else {
    Corelation_DF$WH[i] = 2
  }
  
  if (Corelation_DF$BF[i] == "Every hour") {
    Corelation_DF$BF[i] = 1
  } else if (Corelation_DF$BF[i] == "Every 1-2 hours") {
    Corelation_DF$BF[i] = 1
  } else if (Corelation_DF$BF[i] == "Every 3-4 hours") {
    Corelation_DF$BF[i] = 2
  } else {
    Corelation_DF$BF[i] = 2
  }
  
  
  if (Corelation_DF$NP[i] == "1-2") {
    Corelation_DF$NP[i] = 1
  } else if (Corelation_DF$NP[i] == "3-4") {
    Corelation_DF$NP[i] = 2
  } else if (Corelation_DF$NP[i] == "5-6") {
    Corelation_DF$NP[i] = 3
  } else if (Corelation_DF$NP[i] == "7-9") {
    Corelation_DF$NP[i] = 3
  } else if (Corelation_DF$NP[i] == "") {
    Corelation_DF$NP[i] = NA
  } else {
    Corelation_DF$NP[i] = 3
  }
  
  if (Corelation_DF$SR[i] == "< 10%") {
    Corelation_DF$SR[i] = 1
  } else if (Corelation_DF$SR[i] == "10-20%") {
    Corelation_DF$SR[i] = 2
  } else if (Corelation_DF$SR[i] == "20-30%") {
    Corelation_DF$SR[i] = 3
  } else if (Corelation_DF$SR[i] == "30-50%") {
    Corelation_DF$SR[i] = 4
  } else if (Corelation_DF$SR[i] == "50-75%") {
    Corelation_DF$SR[i] = 5
  } else if (Corelation_DF$SR[i] == "75-90%") {
    Corelation_DF$SR[i] = 6
  } else if (Corelation_DF$SR[i] == "") {
    Corelation_DF$SR[i] = NA
  } else {
    Corelation_DF$SR[i] = 7
  }
  

  
  
  
  if (Corelation_DF$AP[i] == "1-2 weeks") {
    Corelation_DF$AP[i] = 1
  } else if (Corelation_DF$AP[i] == "< 1 week") {
    Corelation_DF$AP[i] = 1
  } else if (Corelation_DF$AP[i] == "2-4 weeks") {
    Corelation_DF$AP[i] = 1
  } else if (Corelation_DF$AP[i] == "1-2 months") {
    Corelation_DF$AP[i] = 2
  } else if (Corelation_DF$AP[i] == "") {
    Corelation_DF$AP[i] = NA
  } else {
    Corelation_DF$AP[i] = 2
  }
  
  if (Corelation_DF$AR[i] == "< 1 month") {
    Corelation_DF$AR[i] = 1
  } else if (Corelation_DF$AR[i] == "1-3 months") {
    Corelation_DF$AR[i] = 2
  } else if (Corelation_DF$AR[i] == "3-6 months") {
    Corelation_DF$AR[i] = 3
  } else if (Corelation_DF$AR[i] == "6-12 months") {
    Corelation_DF$AR[i] = 4
  } else if (Corelation_DF$AR[i] == "") {
    Corelation_DF$AR[i] = NA
  } else {
    Corelation_DF$AR[i] = 5
  }
  
  
  if (Corelation_DF$DWH[i] == "Significantly less") {
    Corelation_DF$DWH[i] = 1
  } else if (Corelation_DF$DWH[i] == "Less") {
    Corelation_DF$DWH[i] = 1
  } else if (Corelation_DF$DWH[i] == "About the same") {
    Corelation_DF$DWH[i] = 1
  } else if (Corelation_DF$DWH[i] == "More") {
    Corelation_DF$DWH[i] = 2
  } else if (Corelation_DF$DWH[i] == "") {
    Corelation_DF$DWH[i] = NA
  } else {
    Corelation_DF$DWH[i] = 2
  }
  
  
  if (Corelation_DF$T[i] == "Minutes before deadline") {
    Corelation_DF$T[i] = 1
  } else if (Corelation_DF$T[i] == "1-3 hours before deadline") {
    Corelation_DF$T[i] = 1
  } else if (Corelation_DF$T[i] == "3-6 hours before deadline") {
    Corelation_DF$T[i] = 1
  } else if (Corelation_DF$T[i] == "1 day before deadline") {
    Corelation_DF$T[i] = 2
  } else if (Corelation_DF$T[i] == "") {
    Corelation_DF$T[i] = NA
  } else {
    Corelation_DF$T[i] = 2
  }
  
  
  if (Corelation_DF$DS[i] == "Extremely less") {
    Corelation_DF$DS[i] = 1
  } else if (Corelation_DF$DS[i] == "Significantly less") {
    # Corelation_DF$DS[i]=2
    Corelation_DF$DS[i] = 1
  } else if (Corelation_DF$DS[i] == "Same") {
    # Corelation_DF$DS[i]=3
    Corelation_DF$DS[i] = 1
  } else if (Corelation_DF$DS[i] == "Significantly  more") {
    # Corelation_DF$DS[i]=4
    Corelation_DF$DS[i] = 2
  } else if (Corelation_DF$DS[i] == "Extremely  more") {
    # Corelation_DF$DS[i]=5
    Corelation_DF$DS[i] = 2
  } else {
    Corelation_DF$DS[i] = NA
  }
  
  if (Corelation_DF$FA[i]=="NSF"){
    Corelation_DF$FA[i]=1
  } else if (Corelation_DF$FA[i]=="NIH"){
    Corelation_DF$FA[i]=2
  } else if (Corelation_DF$FA[i]=="DOE"){
    Corelation_DF$FA[i]=3
  } else if (Corelation_DF$FA[i]=="DOD"){
    Corelation_DF$FA[i]=4
  } else if (Corelation_DF$FA[i]=="NASA"){
    Corelation_DF$FA[i]=5
  } else if (Corelation_DF$FA[i]==""){
    Corelation_DF$FA[i]=NA
  } else{
    Corelation_DF$FA[i]=6
  }
}



Corelation_DF <- Corelation_DF %>%
  mutate(Rank = recode(Rank, "Professor" = 3, "Assistant Professor" = 1, "Associate Professor" = 2))

cols.num <-  c("RS", "WH", "BF", "NP", "SR", "AP", "AR", "DWH", "T", "DS", "FA")
Corelation_DF[cols.num] <-  sapply(Corelation_DF[cols.num], as.numeric)
Corelation_DF$Departmnet <- NULL
Corelation_DF$State <- NULL
Corelation_DF$Disciplines <- NULL
Corelation_DF$State_Geo_location <- NULL

```


```{r, echo=FALSE}
# file_name = 'KeyData.csv'
file_name = 'KeyData.csv'
write.csv(Corelation_DF,
          file.path(curated_data_dir, file_name),
          row.names = FALSE)
```



<!-- ```{r, echo=FALSE} -->
<!-- library(corrplot) -->
<!-- library(superml) -->


<!-- rank_factor <- factor(Corelation_DF$Rank) -->
<!-- Corelation_DF$Rank <- as.numeric(rank_factor) -->

<!-- # Corelation_DF$BR<-factor(Corelation_DF$BR) -->
<!-- M_proposal <- cor(Corelation_DF[, c(1:26)], use = "complete.obs") -->
<!-- res1 <- cor.mtest(Corelation_DF[, c(1:26)], conf.level = .95) -->

<!-- # significance_plot<- -->
<!-- corrplot( -->
<!--   M_proposal, -->
<!--   p.mat = res1$p, -->
<!--   method = "color", -->
<!--   type = "upper", -->
<!--   title = "Psychometrics-core questions correlation on proposals", -->
<!--   sig.level = c(.001, .01, .05), -->
<!--   pch.cex = .9, -->
<!--   diag = FALSE, -->
<!--   insig = "label_sig", -->
<!--   pch.col = "black", -->
<!--   col = colorRampPalette(c("brown1", "white", "dodgerblue"))(200), -->
<!--   tl.col = "black", -->
<!--   mar = c(0, 0, 1, 0) -->
<!-- ) -->
<!-- ``` -->



<!-- ```{r,echo=FALSE} -->

<!-- # Raw_Data <-Raw_Data[,c(4,52:58, 63,68,69)] -->
<!-- # Raw_Data <-Raw_Data[complete.cases(Raw_Data),] -->
<!-- Corelation_DF_Conf<-data.frame(Corelation_DF[,c(1:16)], Raw_Data[,c(4,52:58, 63,68,69)]) -->
<!-- Corelation_DF_Conf<-Corelation_DF_Conf[complete.cases(Corelation_DF_Conf),] -->

<!-- Corelation_DF_Conf<-Corelation_DF_Conf  %>%  dplyr::rename( -->
<!--                                            NCP="wha_average_number", -->
<!--                                            CR="core_rank",                         ## Research Style -->
<!--                                            SR="if_you_submit_manuscripts",    ## Typical week working hours -->
<!--                                            ACP="far_in_advance_do_you",              ## Typical week Research -->
<!--                                            AR="length_of_supp",     ## Break Frequency -->
<!--                                            DWH="in_the_week_leading_to_a_c",      ## Email respnse style -->
<!--                                            DWR="research", -->
<!--                                            DDR="research1", -->
<!--                                            T="you_typically_subm",     ## Submit Proposal -->
<!--                                            DS="ss_level_in_a_fundi"     ## Stress leven on Proposal Deadline -->
<!--                                            ) -->
<!-- ``` -->



<!-- ```{r,echo=FALSE} -->
<!-- for (i in 1:nrow(Corelation_DF_Conf)){ -->
<!--    if (Corelation_DF_Conf$NCP[i]=="1-2"){ -->
<!--     Corelation_DF_Conf$NCP[i]=1 -->
<!--   } else if (Corelation_DF_Conf$NCP[i]=="3-4"){ -->
<!--     Corelation_DF_Conf$NCP[i]=2 -->
<!--   } else if (Corelation_DF_Conf$NCP[i]=="5-6"){ -->
<!--     Corelation_DF_Conf$NCP[i]=3 -->
<!--   } else if (Corelation_DF_Conf$NCP[i]=="7-9"){ -->
<!--     Corelation_DF_Conf$NCP[i]=3 -->
<!--   } else { -->
<!--     Corelation_DF_Conf$NCP[i]=3 -->
<!--   } -->


<!--   if (Corelation_DF_Conf$CR[i]=="C"){ -->
<!--     Corelation_DF_Conf$CR[i]=1 -->
<!--   } else if (Corelation_DF_Conf$CR[i]=="B"){ -->
<!--     Corelation_DF_Conf$CR[i]=2 -->
<!--   }  else if (Corelation_DF_Conf$CR[i]=="A"){ -->
<!--     Corelation_DF_Conf$CR[i]=3 -->
<!--   } else { -->
<!--     Corelation_DF_Conf$CR[i]=3 -->
<!--   } -->

<!--   if (Corelation_DF_Conf$SR[i]=="< 10%"){ -->
<!--     Corelation_DF_Conf$SR[i]=0 -->
<!--   } else if (Corelation_DF_Conf$SR[i]=="10-20%"){ -->
<!--     Corelation_DF_Conf$SR[i]=0 -->
<!--   } else if (Corelation_DF_Conf$SR[i]=="20-30%"){ -->
<!--     Corelation_DF_Conf$SR[i]=0 -->
<!--   } else if (Corelation_DF_Conf$SR[i]=="30-50%"){ -->
<!--     Corelation_DF_Conf$SR[i]=0 -->
<!--   } else if (Corelation_DF_Conf$SR[i]=="50-75%"){ -->
<!--     Corelation_DF_Conf$SR[i]=1 -->
<!--   } else if (Corelation_DF_Conf$SR[i]=="75-90%"){ -->
<!--     Corelation_DF_Conf$SR[i]=1 -->
<!--   } else{ -->
<!--     Corelation_DF_Conf$SR[i]=1 -->
<!--   } -->

<!--   if (Corelation_DF_Conf$ACP[i]=="< 1 week"){ -->
<!--     Corelation_DF_Conf$ACP[i]=1 -->
<!--   } else if (Corelation_DF_Conf$ACP[i]=="1-2 weeks"){ -->
<!--     Corelation_DF_Conf$ACP[i]=1 -->
<!--   } else if (Corelation_DF_Conf$ACP[i]=="2-4 weeks"){ -->
<!--     Corelation_DF_Conf$ACP[i]=1 -->
<!--   } else if (Corelation_DF_Conf$ACP[i]=="1-2 months"){ -->
<!--     Corelation_DF_Conf$ACP[i]=2 -->
<!--   } else { -->
<!--     Corelation_DF_Conf$ACP[i]=3 -->
<!--   } -->

<!--   if (Corelation_DF_Conf$AR[i]=="< 1 month"){ -->
<!--     Corelation_DF_Conf$AR[i]=1 -->
<!--   } else if(Corelation_DF_Conf$AR[i]=="1-3 months"){ -->
<!--     Corelation_DF_Conf$AR[i]=1 -->
<!--   } else if(Corelation_DF_Conf$AR[i]=="3-6 months"){ -->
<!--     Corelation_DF_Conf$AR[i]=1 -->
<!--   } else if(Corelation_DF_Conf$AR[i]=="6-12 months"){ -->
<!--     Corelation_DF_Conf$AR[i]=2 -->
<!--   } else if(Corelation_DF_Conf$AR[i]==""){ -->
<!--     Corelation_DF_Conf$AR[i]=NA -->
<!--   } else { -->
<!--     Corelation_DF_Conf$AR[i]=3 -->
<!--   } -->


<!--   if (Corelation_DF_Conf$DWH[i]=="Significantly less"){ -->
<!--     Corelation_DF_Conf$DWH[i]=1 -->
<!--   } else if (Corelation_DF_Conf$DWH[i]=="Less"){ -->
<!--     Corelation_DF_Conf$DWH[i]=1 -->
<!--   } else if (Corelation_DF_Conf$DWH[i]=="About the same"){ -->
<!--     Corelation_DF_Conf$DWH[i]=1 -->
<!--   } else if (Corelation_DF_Conf$DWH[i]=="More"){ -->
<!--     Corelation_DF_Conf$DWH[i]=2 -->
<!--   } else if (Corelation_DF_Conf$DWH[i]==""){ -->
<!--     Corelation_DF_Conf$DWH[i]=NA -->
<!--   } else { -->
<!--     Corelation_DF_Conf$DWH[i]=3 -->
<!--   } -->


<!--   if (Corelation_DF_Conf$T[i]=="Minutes before deadline"){ -->
<!--     Corelation_DF_Conf$T[i]= 1 -->
<!--   } else if (Corelation_DF_Conf$T[i]=="1-3 hours before deadline"){ -->
<!--     Corelation_DF_Conf$T[i]=2 -->
<!--   } else if (Corelation_DF_Conf$T[i]=="3-6 hours before deadline"){ -->
<!--     Corelation_DF_Conf$T[i]=3 -->
<!--   } else if (Corelation_DF_Conf$T[i]=="1 day before deadline"){ -->
<!--     Corelation_DF_Conf$T[i]=4 -->
<!--   } else if (Corelation_DF_Conf$T[i]==""){ -->
<!--     Corelation_DF_Conf$T[i]=NA -->
<!--   } else { -->
<!--     Corelation_DF_Conf$T[i]=5 -->
<!--   } -->


<!--   if (Corelation_DF_Conf$DS[i]=="Extremely less"){ -->
<!--     Corelation_DF_Conf$DS[i]=1 -->
<!--   } else if (Corelation_DF_Conf$DS[i]=="Significantly less"){ -->
<!--     Corelation_DF_Conf$DS[i]=1 -->
<!--   } else if (Corelation_DF_Conf$DS[i]=="Same"){ -->
<!--     Corelation_DF_Conf$DS[i]=1 -->
<!--   } else if (Corelation_DF_Conf$DS[i]=="Significantly  more"){ -->
<!--     Corelation_DF_Conf$DS[i]=2 -->
<!--   } else if (Corelation_DF_Conf$DS[i]=="Extremely  more"){ -->
<!--     Corelation_DF_Conf$DS[i]=3 -->
<!--   } else { -->
<!--     Corelation_DF_Conf$DS[i]=NA -->
<!--   } -->
<!-- } -->


<!-- # colnames(Corelation_DF_Conf) -->

<!-- cols.num <- c("NCP", "CR", "SR", "ACP", "AR", "DWH", "DWR", "DDR", "T", "DS") -->
<!-- Corelation_DF_Conf[cols.num] <- sapply(Corelation_DF_Conf[cols.num],as.numeric) -->

<!-- Conf_C_Eng<-Corelation_DF_Conf -->
<!-- Conf_A_Astar<-Corelation_DF_Conf -->
<!-- Conf_B_C<-Corelation_DF_Conf -->

<!-- Corelation_DF_Conf$departmental_affiliation<-NULL -->
<!-- ``` -->

<!-- ```{r, echo=FALSE} -->
<!-- file_name = 'CorrelationData_conf_SR2.csv' -->
<!-- write.csv(Corelation_DF_Conf, -->
<!--           file.path(curated_data_dir, file_name), -->
<!--           row.names = FALSE) -->
<!-- ``` -->


<!-- ```{r,echo=FALSE} -->
<!-- #  -->
<!-- # # Defines function to color according to correlation -->
<!-- # cor_func <- function(data, mapping, method, symbol, ...) { -->
<!-- #   x <- eval_data_col(data, mapping$x) -->
<!-- #   y <- eval_data_col(data, mapping$y) -->
<!-- #   corr <- cor(x, y, method = method, use = 'complete.obs') -->
<!-- #   colFn <- colorRampPalette(c("brown1", "white", "dodgerblue"), -->
<!-- #                             interpolate = 'spline') -->
<!-- #   fill <- colFn(100)[findInterval(corr, seq(-1, 1, length = 100))] -->
<!-- #    -->
<!-- #   ggally_text( -->
<!-- #     label = paste(symbol, as.character(round(corr, 2))), -->
<!-- #     mapping = aes(), -->
<!-- #     xP = 0.5, -->
<!-- #     yP = 0.5, -->
<!-- #     color = 'black', -->
<!-- #     size = 2.75, -->
<!-- #     ... -->
<!-- #   ) + -->
<!-- #     theme(panel.background = element_rect(fill = fill)) -->
<!-- # } -->
<!-- #  -->
<!-- # pm <- ggpairs( -->
<!-- #   Corelation_DF_Conf, -->
<!-- #   upper = list(continuous = wrap( -->
<!-- #     cor_func, -->
<!-- #     # method = 'spearman', symbol = expression('\u03C1 ='))), -->
<!-- #     method = 'spearman', symbol = expression('') -->
<!-- #   )), -->
<!-- #    -->
<!-- #   lower = list( -->
<!-- #     continuous = function(data, mapping, ...) { -->
<!-- #       ggally_smooth_lm(data = data, mapping = mapping) + -->
<!-- #         theme(panel.background = element_blank()) -->
<!-- #     } -->
<!-- #   ), -->
<!-- #   diag = list( -->
<!-- #     continuous = function(data, mapping, ...) { -->
<!-- #       ggally_densityDiag(data = data, mapping = mapping) + -->
<!-- #         theme(panel.background = element_blank()) -->
<!-- #     } -->
<!-- #   ) -->
<!-- # ) -->
<!-- #  -->
<!-- # mytheme = theme( -->
<!-- #   strip.background = element_rect(fill = "white"), -->
<!-- #   panel.grid.major = element_blank(), -->
<!-- #   panel.grid.minor = element_blank() -->
<!-- # ) -->
<!-- #  -->
<!-- #  -->
<!-- # Cor_plot<-pm + mytheme -->
<!-- #  -->
<!-- #       filename<-"Conf_Corelation.pdf" -->
<!-- #       full_path<-file.path(plot_dir, filename) -->
<!-- #       ggsave(full_path, Cor_plot, width = 11, height = 11, units = "in") -->
<!-- ``` -->



<!-- ```{r,echo=FALSE} -->
<!-- library(corrplot) -->
<!-- M_Conf <- cor(Corelation_DF_Conf[, c(2:26)], use = "complete.obs") -->
<!-- res2 <- cor.mtest(Corelation_DF_Conf[, c(2:26)], conf.level = .95) -->

<!-- corrplot( -->
<!--   M_Conf, -->
<!--   p.mat = res2$p, -->
<!--   method = "color", -->
<!--   type = "upper", -->
<!--   # title = "Psychometrics-core questions correlation on conference papers", -->
<!--   title = "Psychometrics-core questions correlation on conference papers", -->
<!--   sig.level = c(.001, .01, .05), -->
<!--   pch.cex = .9, -->
<!--   diag = FALSE, -->
<!--   insig = "label_sig", -->
<!--   pch.col = "black", -->
<!--   col = colorRampPalette(c("brown1", "white", "dodgerblue"))(200), -->
<!--   tl.col = "black", -->
<!--   mar = c(0, 0, 1, 0) -->
<!-- ) -->

<!-- ``` -->


<!-- ```{r, echo=FALSE} -->
<!-- Conf_C_Eng <- -->
<!--   subset( -->
<!--     Conf_C_Eng, -->
<!--     departmental_affiliation == "Computer Science" | -->
<!--       departmental_affiliation == "Engineering" -->
<!--   ) -->
<!-- Conf_C_Eng$departmental_affiliation <- NULL -->


<!-- M_Conf_c_eng <- cor(Conf_C_Eng[, c(2:26)], use = "complete.obs") -->
<!-- res3 <- cor.mtest(Conf_C_Eng[, c(2:26)], conf.level = .95) -->

<!-- # significance_plot<- -->

<!-- # par(mfrow = c(2, 2)) -->
<!-- corrplot( -->
<!--   M_Conf_c_eng, -->
<!--   p.mat = res3$p, -->
<!--   method = "color", -->
<!--   type = "upper", -->
<!--   title = "Psychometrics-core questions correlation on conference papers Com-Eng", -->
<!--   sig.level = c(.001, .01, .05), -->
<!--   pch.cex = .9, -->
<!--   diag = FALSE, -->
<!--   insig = "label_sig", -->
<!--   pch.col = "black", -->
<!--   col = colorRampPalette(c("brown1", "white", "dodgerblue"))(200), -->
<!--   tl.col = "black", -->
<!--   mar = c(0, 0, 1, 0) -->
<!-- ) -->
<!-- ``` -->

<!-- ```{r, echo=FALSE} -->
<!-- Conf_A_Astar$departmental_affiliation <- NULL -->
<!-- Conf_A_Astar <- subset(Conf_A_Astar, CR == "3" | CR == "4") -->

<!-- M_Conf_A_Astar <- cor(Conf_A_Astar[, c(2:26)], use = "complete.obs") -->
<!-- res4 <- cor.mtest(Conf_A_Astar[, c(2:26)], conf.level = .95) -->

<!-- corrplot( -->
<!--   M_Conf_A_Astar, -->
<!--   p.mat = res4$p, -->
<!--   method = "color", -->
<!--   type = "upper", -->
<!--   title = "Psychometrics-core questions correlation on A-A* conference papers", -->
<!--   sig.level = c(.001, .01, .05), -->
<!--   pch.cex = .9, -->
<!--   diag = FALSE, -->
<!--   insig = "label_sig", -->
<!--   pch.col = "black", -->
<!--   col = colorRampPalette(c("brown1", "white", "dodgerblue"))(200), -->
<!--   tl.col = "black", -->
<!--   mar = c(0, 0, 1, 0) -->
<!-- ) -->
<!-- # dev.off() -->
<!-- ``` -->


<!-- ```{r, echo=FALSE} -->
<!-- Conf_B_C$departmental_affiliation <- NULL -->
<!-- Conf_B_C <- subset(Conf_B_C, CR == "1" | CR == "2") -->

<!-- M_Conf_B_C <- cor(Conf_B_C[, c(2:26)], use = "complete.obs") -->
<!-- res5 <- cor.mtest(Conf_B_C[, c(2:26)], conf.level = .95) -->

<!-- # significance_plot<- -->
<!-- # pdf(file = "corplot.pdf") -->
<!-- corrplot( -->
<!--   M_Conf_B_C, -->
<!--   p.mat = res5$p, -->
<!--   method = "color", -->
<!--   type = "upper", -->
<!--   title = "Psychometrics-core questions correlation on B-C conference papers", -->
<!--   sig.level = c(.001, .01, .05), -->
<!--   pch.cex = .9, -->
<!--   diag = FALSE, -->
<!--   insig = "label_sig", -->
<!--   pch.col = "black", -->
<!--   col = colorRampPalette(c("brown1", "white", "dodgerblue"))(200), -->
<!--   tl.col = "black", -->
<!--   mar = c(0, 0, 1, 0) -->
<!-- ) -->
<!-- # dev.off() -->
<!-- ``` -->


<!-- ```{r, echo=FALSE} -->

<!-- par(mfrow = c(2, 2)) -->

<!-- corrplot( -->
<!--   M_proposal, -->
<!--   p.mat = res1$p, -->
<!--   method = "color", -->
<!--   type = "upper", -->
<!--   title = "Proposals", -->
<!--   sig.level = c(.001, .01, .05), -->
<!--   pch.cex = .9, -->
<!--   diag = FALSE, -->
<!--   insig = "label_sig", -->
<!--   pch.col = "black", -->
<!--   col = colorRampPalette(c("brown1", "white", "dodgerblue"))(200), -->
<!--   tl.col = "black", -->
<!--   mar = c(0, 0, 1, 0) -->
<!-- ) -->
<!-- corrplot( -->
<!--   M_Conf, -->
<!--   p.mat = res2$p, -->
<!--   method = "color", -->
<!--   type = "upper", -->
<!--   title = "Conference papers", -->
<!--   sig.level = c(.001, .01, .05), -->
<!--   pch.cex = .9, -->
<!--   diag = FALSE, -->
<!--   insig = "label_sig", -->
<!--   pch.col = "black", -->
<!--   col = colorRampPalette(c("brown1", "white", "dodgerblue"))(200), -->
<!--   tl.col = "black", -->
<!--   mar = c(0, 0, 1, 0) -->
<!-- ) -->
<!-- corrplot( -->
<!--   M_Conf_A_Astar, -->
<!--   p.mat = res4$p, -->
<!--   method = "color", -->
<!--   type = "upper", -->
<!--   title = "A-A* conference papers", -->
<!--   sig.level = c(.001, .01, .05), -->
<!--   pch.cex = .9, -->
<!--   diag = FALSE, -->
<!--   insig = "label_sig", -->
<!--   pch.col = "black", -->
<!--   col = colorRampPalette(c("brown1", "white", "dodgerblue"))(200), -->
<!--   tl.col = "black", -->
<!--   mar = c(0, 0, 1, 0) -->
<!-- ) -->
<!-- corrplot( -->
<!--   M_Conf_B_C, -->
<!--   p.mat = res5$p, -->
<!--   method = "color", -->
<!--   type = "upper", -->
<!--   title = "B-C conference papers", -->
<!--   sig.level = c(.001, .01, .05), -->
<!--   pch.cex = .9, -->
<!--   diag = FALSE, -->
<!--   insig = "label_sig", -->
<!--   pch.col = "black", -->
<!--   col = colorRampPalette(c("brown1", "white", "dodgerblue"))(200), -->
<!--   tl.col = "black", -->
<!--   mar = c(0, 0, 1, 0) -->
<!-- ) -->

<!-- ``` -->


<!-- ```{r, echo=FALSE} -->

<!-- par(mfrow = c(2, 1)) -->
<!-- corrplot( -->
<!--   M_Conf, -->
<!--   p.mat = res2$p, -->
<!--   method = "color", -->
<!--   type = "upper", -->
<!--   title = "Psychometrics-core questions correlation on conference papers", -->
<!--   sig.level = c(.001, .01, .05), -->
<!--   pch.cex = .9, -->
<!--   diag = FALSE, -->
<!--   insig = "label_sig", -->
<!--   pch.col = "black", -->
<!--   col = colorRampPalette(c("brown1", "white", "dodgerblue"))(200), -->
<!--   tl.col = "black", -->
<!--   mar = c(0, 0, 1, 0) -->
<!-- ) -->
<!-- corrplot( -->
<!--   M_Conf_c_eng, -->
<!--   p.mat = res3$p, -->
<!--   method = "color", -->
<!--   type = "upper", -->
<!--   title = "Psychometrics-core questions correlation on conference papers Com-Eng", -->
<!--   sig.level = c(.001, .01, .05), -->
<!--   pch.cex = .9, -->
<!--   diag = FALSE, -->
<!--   insig = "label_sig", -->
<!--   pch.col = "black", -->
<!--   col = colorRampPalette(c("brown1", "white", "dodgerblue"))(200), -->
<!--   tl.col = "black", -->
<!--   mar = c(0, 0, 1, 0) -->
<!-- ) -->
<!-- ``` -->
