==> scripts/5-vs-descriptive-statistics.rmd <==
---
title: "Descriptive Statistics Visulaization (Figure1 and Figure5 )"
output:
  pdf_document: default
  html_document: default
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(freqtables)
library(tools)
library(plotrix)
library(plotly)
library(grid)
library(dplyr)
library(gridExtra)
library(ggpubr)
library(cowplot)
library(reshape2)
library(plyr)
library(ggplot2)
library(ggrepel)
library(forcats)
library(scales)
library(stringr)
```


```{r, echo=FALSE, warning=FALSE}
root_dir <- getwd()
project_directory <- dirname(root_dir)
data_dir <- file.path(project_directory, 'curated-data')
plot_dir <- file.path(project_directory, 'Plots')
# data_file_name <- 'Selected_coreQuestions.csv'
data_file_name <- 'CoreQuestions.csv'
curated_data_dir <- file.path(project_directory, 'curated-data')
Data <-
  read.csv(file.path(data_dir, data_file_name), stringsAsFactors = FALSE)
```

```{r, echo=FALSE, warning=FALSE}
Data <- Data[,c("FR", "Gender", "Departmnet", "State", "TW_W_H", "TW_Research", "RO_F_EG")]

Data <- Data %>%
  mutate(FR = recode(FR, "Assistant Professor" = "FR1", "Associate Professor" = "FR2", "Professor" = "FR3" ))
Data$TW_Research <- as.factor(Data$TW_Research)
Data <- Data %>%
  mutate(TWR = recode(TW_Research, "0" = "0-20", "10" = "0-20", "20" = "0-20", "70" = "70-90", "80" = "70-90",  "90" = "70-90")) %>% 
  mutate(EG = recode(RO_F_EG, "1-25%" = "1-25", "25-50%" = "25-50", "50-75%" = "50-75", "75-100%" = "75-100", "Fully funded" = "FF",  "Not funded" = "NF"))

```


```{r, echo=FALSE, warning=FALSE}
Natural_Sciences = '^Applied Math$|^Chemistry and Biochemistry$|^Oceanography$|^Environmental Studies$|^Chemistry$|^Physics$|^Geosciences$|^Mathematics$'
CIS = '^Informatics$|^Information$|^Information and Logistics Technology$|^Information Science$|^Computer Science$|^Scientific Computing$'
Engineering = '^Engineering / Learning Sciences$|^Management$|^Statistical Sciences and Operations Research$|^Technology$|^Engineering$'
Biomedical_Sciences = '^Health$|^Health and Human Performance$|^HHP$|^Neurobiology$|^Neuroscience$|^Optometry$|^Pharmacological and Pharmaceutical Sciences$|^Pharmacy$|^Vision Science$|^Biology$|^Medicine$|^Pharmacy $'
Behavioral_Sciences = '^Political Science$|^Sociology$|^Psychology$|^Speech, Language, and Hearing$|^Political Science $|^Communication Studies$'



Data <- Data %>%
  mutate(
    Disciplines = case_when(
      # str_detect(Departmnet, Natural_Sciences) ~ 'Natural Sciences',
      str_detect(Departmnet, Natural_Sciences) ~ 'NAT',
      # str_detect(Departmnet, CIS) ~ 'Computer & Information Sciences',
      str_detect(Departmnet, CIS) ~ 'CIS',
      # str_detect(Departmnet, Engineering) ~ 'Engineering',
      str_detect(Departmnet, Engineering) ~ 'ENG',
      # str_detect(Departmnet, Biomedical_Sciences) ~ 'Biomedical Sciences',
      str_detect(Departmnet, Biomedical_Sciences) ~ 'BIO',
      # str_detect(Departmnet, Behavioral_Sciences) ~ 'Behavioral Sciences'
      str_detect(Departmnet, Behavioral_Sciences) ~ 'BEHAV'
    )
  )

```



```{r, echo=FALSE, warning=FALSE}
core_plot_df <- Data[,c("Gender", "Disciplines", "State", "TW_W_H", "TWR", "EG")]
core_plot_df <- core_plot_df %>%
  mutate(TWR = recode(TWR, "0-20" = "< 20", "70-90" = "> 70", "100" = "> 70"))
```



```{r, echo=FALSE, warning=FALSE}
list_gender = c("Female", "Male")
list0 = c("BIO", "BEHAV", "ENG", "CIS", "NAT")
list1 = c("East", "West", "Midwest", "South")
list2 = c("< 30", "30-40", "40-50", "> 50")
list3 = c("< 20", "30", "40", "50", "60", "> 70")
list4 = c("NF",
          "1-25",
          "25-50",
          "50-75",
          "75-100",
          "FF")
Ticks_list <-
  list(list_gender, 
       list0,
       list1,
       list2,
       list3,
       list4)
title_list <- c("Gender Distribution", "Disciplinary Distribution", "Geographic Distribution", "Weekly Workload [hrs]",  "Research Load [%]", "Funding Coverage [%]")
```



```{r, echo=FALSE, warning=FALSE}
percent <- function(x, digits = 0, format = "f", ...) {
  paste0(formatC(100 * x, format = format, digits = digits, ...), "%")
}
```


```{r, echo=FALSE, warning=FALSE}

plot_list <- list()

for (i in 1:length(core_plot_df)) {
  y_title = ""
  if (title_list[i] == "Weekly Workload [hrs]" | title_list[i] == "Gender Distribution") {
    y_title = "# of Respondents"
  }
  
  # ylimit  = 120


  if (i == 1) {
    ylimit = 300
  }  else if (i == 2 | i == 3) {
    ylimit = 200
  }  else if (i == 4) {
    ylimit = 250
  } else {
    ylimit = 105
  }


  temp <- count(core_plot_df[i])
  colnames(temp) <- c("item", "count")
  # temp <- temp[complete.cases(temp), ]
  temp <- temp[!(temp$item == ""),]
  temp <- temp[order(temp$count), ]
  temp <- temp %>% mutate(prop = (count / sum(count))) 
  temp$percentage <- percent((temp$prop))
  
  if (i == 1) {
    temp$percentage[2] <- "67.5%"
    temp$percentage[1] <- "32.5%"
     }

  if (i == 3) {
    temp$percentage[4] <- "43%"
  }
  if (i == 4) {
    temp$percentage[3] <- "35%"
  }
  if (i == 6) {
    temp$percentage[6] <- "26%"
  }
  

  bar_plot <- ggplot(data = temp, aes(x = item, y = count)) +
    geom_bar(stat = "identity",
             width = 0.5,
             fill = "steelblue") +
    geom_text(
      aes(label = percentage),
      # vjust = 0.5,
      vjust = -0.25,
      color = "black",
      # angle = 90,
      size = 2.5,
      face = "bold"
    ) +
    theme_bw() +
    scale_y_continuous(breaks = seq(0, ylimit, by = 50),
                       limits = c(0, ylimit)) +
    labs(x = "", y = y_title, title = title_list[i]) +
    theme(
      axis.title.y = element_text(
        size = 8,
        face = "bold",
        colour = "black"
      ),
      panel.grid.major = element_blank(),
      panel.grid.minor = element_blank(),
      plot.title = element_text(hjust = 0.5, size = 10),
      axis.text.x = element_text(face = "bold", size = 7),
      axis.text.y = element_text(face = "bold", size = 7)
    ) + scale_x_discrete(limits = Ticks_list[[i]])
  # print(bar_plot)

  plot_list[[length(plot_list) + 1]] <- bar_plot
  

}


final_plot <- ggarrange(plotlist = plot_list,
                        nrow = 2,
                        ncol = 3)
filename <- "Figure1.pdf"
full_path <- file.path(plot_dir, filename)
ggsave(
  full_path,
  final_plot,
  width = 8.5,
  height = 5.5,
  units = "in"
)

```


```{r, echo=FALSE, warning=FALSE}
# root_dir <- getwd()
# project_directory <- dirname(root_dir)
# data_dir <- file.path(project_directory, 'curated-data')
# plot_dir <- file.path(project_directory, 'Plots')
data_file_name <- 'AllData.csv'
curated_data_dir <- file.path(project_directory, 'curated-data')
Data <- read.csv(file.path(data_dir, data_file_name), stringsAsFactors = FALSE)
# colnames(Data)
Data <- Data[,c("P_Disrupted_Research", "P_Disrupted_Sleep", "P_Disrupted_Diet", "P_Disrupted_PA", "P_Disrupted_IR")]

colnames(Data) <- c("RR", "Sleep", "Diet", "Physical Activity", "Interpersonal Relationships")



Data_Melt <-
  melt(Data[, c("RR",
                "Sleep",
                "Diet",
                "Physical Activity",
                "Interpersonal Relationships")])

Data_Melt_count <- Data_Melt %>% 
  group_by(variable, value)  %>% 
  dplyr::summarise(n = n())

Data_Melt_count <- Data_Melt_count[complete.cases(Data_Melt_count), ]
Data_Melt_count <- Data_Melt_count %>%
    group_by(variable) %>%
    mutate(countT= sum(n)) %>%
    group_by(variable, add=TRUE) %>%
    mutate(per=paste0(round(100*n/countT,2)))
Data_Melt_count$value <- as.factor(Data_Melt_count$value)
Data_Melt_count$per <- as.numeric(Data_Melt_count$per)

```




```{r, echo=FALSE, warning=FALSE}
library(ggplot2)

figure5 <-
  ggplot(Data_Melt_count, aes(fill = value, y = per, x = variable)) +
  geom_bar(stat = "identity",
           position = position_stack(reverse = TRUE),
           colour = "black") +
  scale_fill_manual("No Disruption",values = c("#0000FF", "#e0ffff", "#FFFFFF", "#FFC0CB", "#FF0000"), labels = c("", "", "", "", "Extreme Disruption")) +
  theme_bw() +
  theme(
    panel.grid = element_blank(),
    legend.position = "top",
    legend.text = element_text(face = "bold", color = "black", size = 9),
    legend.title = element_text(face = "bold", color = "black", size = 9),
    axis.text.x = element_text(face = "bold", size = 9,color = "black"),
    axis.text.y = element_text(face = "bold", size = 9,color = "black"),
    axis.title.y = element_text(face = "bold", size = 9,color = "black")
  ) +
  labs(x = element_blank(),
       y = "Respondents [ % ]") +
  scale_x_discrete(
    limits = c(
      "Diet",
      "Interpersonal Relationships",
      "Physical Activity",
      "Sleep",
      "RR"
    ),
    labels = c(
      "Diet",
      "Interpersonal\n       Relationships",
      "Physical\n       Activity",
      "Sleep",
      "Regular\n             Research"
    )
  ) 

figure5
filename <- "Figure5.pdf"
full_path <- file.path(plot_dir, filename)
ggsave(
  full_path,
  figure5,
  width = 8.5,
  height = 5.5,
  units = "in"
)
```
