==> scripts/6-vs-model-grantsmanship.Rmd <==
---
title: "Grantsmanship Models"
output:
  html_document:
    df_print: paged
---

```{r setup, include=FALSE}
# knitr::opts_chunk$set(echo = TRUE)
knitr::opts_chunk$set(echo = TRUE)
library(dplyr)
library(ggplot2)
library(GGally)
library(reshape2)
library(plyr)
library(nnet)
library(MASS)
library(caret)
library(mlbench)
library(rms)
require(tidyr)
library(superml)
library(randomForest)
library(aod)
library(pROC)
library(interplot)
library(sjPlot)
library(grid)
library(gridExtra)
library(ggpubr)
library(freqtables)
library(cowplot)
```


```{r, echo=FALSE, warning=FALSE}
  root_dir <- getwd()
  project_directory <- dirname(root_dir)
  data_dir <- file.path(project_directory, 'raw-data')
  curated_data_dir <- file.path(project_directory, 'curated-data')
  plot_dir <- file.path(project_directory, 'Plots')
```


```{r, echo=FALSE, warning=FALSE}

data_processing_model <- function(data_file_name) {
  root_dir <- getwd()
  project_directory <- dirname(root_dir)
  data_dir <- file.path(project_directory, 'raw-data')
  curated_data_dir <- file.path(project_directory, 'curated-data')
  plot_dir <- file.path(project_directory, 'Plots')
  
  # print(data_file_name)
  lm_DF <-
    read.csv(file.path(curated_data_dir, data_file_name),
             stringsAsFactors = FALSE)
  # count(lm_DF$SR)
  
  lm_DF<- lm_DF  %>%  dplyr::rename(AC = "AV")
  
  lm_DF$RS <- as.factor(lm_DF$RS)
  lm_DF$WH <- as.factor(lm_DF$WH)
  lm_DF$NP <- as.factor(lm_DF$NP)
  lm_DF$BF <- as.factor(lm_DF$BF)
  lm_DF$DWH <- as.factor(lm_DF$DWH)
  lm_DF$FA <- as.factor(lm_DF$FA)
  lm_DF$TS <- as.factor(lm_DF$TS)
  lm_DF$DS <- as.factor(lm_DF$DS)
  lm_DF$PR <- as.factor(lm_DF$PR)
  lm_DF$FR <- as.factor(lm_DF$FR)
  
  
  
  NullModel <- glm(SR ~ 1, data = lm_DF, family = "binomial")
  FullModel <- glm(SR ~ ., data = lm_DF, family = "binomial")
  BEmodel <- stepAIC(FullModel, direction = "backward", trace = FALSE)
  
  
  
  if (data_file_name == "ModelData_SR20.csv") {
    model_20 <-
      glm(SR ~ H + NP + FA + DS, family = "binomial", data = lm_DF)
    return(model_20)
  } else if (data_file_name == "ModelData_SR30.csv") {
    model_30 <-
      glm(SR ~ NP + DS + H + FA + TA, family = "binomial", data = lm_DF)
    return(model_30)
  } else{
    model_50 <-
      glm(SR ~ NP + FA + H + E + BF + DS + AC + PR,
          family = "binomial",
          data = lm_DF)
    return(model_50)
  }
  
}

```


```{r, echo=FALSE, warning=FALSE}
model_20 <- data_processing_model("ModelData_SR20.csv")
# summary(model_20)


```


```{r, echo=FALSE, warning=FALSE}

Zcrit<-1.96   ### critical value for 95% confidence intervals
  
ilink <- family(model_20)$linkinv
NP<-c("1","2","3",rep("1",9))
FA<-c(rep("1",3),"1","2","3","4","5","6",rep("1",3))
DS<-c(rep("1",9),"1","2", rep("1",1))
H<-rep(29.56824,12)
tmpd<-data.frame(NP,FA,DS,H)
ndata <- cbind(tmpd, setNames(as_tibble(predict(model_20, tmpd, se.fit = TRUE)[1:2]),
                                   c('fit_link','se_link')))
Model_Data_20 <- mutate(ndata,
                fit_resp  = ilink(fit_link),
                Lower_limit = ilink(fit_link - (Zcrit * se_link)),
                Upper_limit = ilink(fit_link + (Zcrit * se_link)))
# Model_Data_20
```


## Successful Grantsmanship
```{r, echo=FALSE, warning=FALSE}
model_30 <- data_processing_model("ModelData_SR30.csv")
summary(model_30)
```


```{r, echo=FALSE, warning=FALSE}

Zcrit<-1.96   ### critical value for 95% confidence intervals
  
ilink <- family(model_30)$linkinv
NP<-c("1","2","3",rep("1",10))
FA<-c(rep("1",3),"1","2","3","4","5","6",rep("1",4))
DS<-c(rep("1",9),"1","2", rep("1",2))
H<-rep(29.56824,13)
TA<-rep(41.33747,13)
tmpd<-data.frame(NP,FA,DS,H,TA)
ndata <- cbind(tmpd, setNames(as_tibble(predict(model_30, tmpd, se.fit = TRUE)[1:2]),
                                   c('fit_link','se_link')))
Model_Data_30 <- mutate(ndata,
                fit_resp  = ilink(fit_link),
                Lower_limit = ilink(fit_link - (Zcrit * se_link)),
                Upper_limit = ilink(fit_link + (Zcrit * se_link)))
# Model_Data_30

# summary(model_30)
```
\newpage
## Highly Successful Grantsmanship
```{r, echo=FALSE, warning=FALSE}
model_50 <- data_processing_model("ModelData_SR50.csv")
summary(model_50)
```

```{r, echo=FALSE, warning=FALSE}
ilink <- family(model_50)$linkinv
NP<-as.factor(c("1","2","3",rep("1",18)))
FA<-as.factor(c(rep("1",3),"1","2","3","4","5","6",rep("1",12)))
BF<-as.factor(c(rep("1",9),"1","2",rep("1",10)))
DS<-as.factor(c(rep("1",11),"1","2",rep("1",8)))
PR<-as.factor(c(rep("1",13),"1","2","3","4","5", rep("1",3)))
H<-rep(29.56824,21)
E<-rep(5.957816,21)
AC<-rep(18.81638,21)
tmpd<-data.frame(NP,FA,BF,DS,PR,H,E,AC)
ndata <- cbind(tmpd, setNames(as_tibble(predict(model_50, tmpd, se.fit = TRUE)[1:2]),
                              c('fit_link','se_link')))
Model_Data_50 <- mutate(ndata,
                fit_resp  = ilink(fit_link),
                Lower_limit = ilink(fit_link - (Zcrit * se_link)),
                Upper_limit = ilink(fit_link + (Zcrit * se_link)))
# Model_Data_50
```



```{r, echo=FALSE, warning=FALSE}
slice_df <- function(DF, R1, R2, Item, Conf){
  sliced_data = DF %>% 
        dplyr::slice(R1:R2) %>% 
        dplyr::select(Item, fit_resp, Lower_limit, Upper_limit) %>%
        dplyr::mutate(Features = Item) %>% 
        dplyr::mutate(Conf = Conf) %>% 
        dplyr::rename("Levels" = Item)
  return(sliced_data)
}
      
Model_Data <- slice_df(Model_Data_20, 1, 3, "NP", "20")
Model_Data <- rbind(Model_Data, slice_df(Model_Data_20, 4, 9, "FA", "20"))
Model_Data <- rbind(Model_Data, slice_df(Model_Data_20, 10, 11, "DS", "20"))
# Model_Data <- rbind(Model_Data, slice_df(Model_Data_20, 10, 11, "TS", "20"))

Model_Data <- rbind(Model_Data, slice_df(Model_Data_30, 1, 3, "NP", "30"))
Model_Data <- rbind(Model_Data, slice_df(Model_Data_30, 4, 9, "FA", "30"))
Model_Data <- rbind(Model_Data, slice_df(Model_Data_30, 10, 11, "DS", "30"))
# Model_Data <- rbind(Model_Data, slice_df(Model_Data_30, 10, 12, "FR", "30"))
# Model_Data <- rbind(Model_Data, slice_df(Model_Data_30, 13, 14, "RS", "30"))


Model_Data <- rbind(Model_Data, slice_df(Model_Data_50, 1, 3, "NP", "50"))
Model_Data <- rbind(Model_Data, slice_df(Model_Data_50, 4, 9, "FA", "50"))
Model_Data <- rbind(Model_Data, slice_df(Model_Data_50, 10, 11, "BF", "50"))
Model_Data <- rbind(Model_Data, slice_df(Model_Data_50, 14, 18, "PR", "50"))


```


```{r, echo=FALSE}
# np_actual = c("1" = "NP1", "2" = "NP2", "3"="NP3")
# fa_actual = c("1" = "NSF", "2" = "NIH", "3" = "DOE", "4" = "DOD", "5" = "NASA", "6" = "OT")
# TS_actual = c("1" = "TS1", "2" = "TS2")
# DS_actual = c("1" = "DS1", "2" = "DS2")
# BF_actual = c("1" = "BF1", "2" = "BF2")
# PR_actual = c("1" = "PR1", "2" = "PR2", "3" = "PR3", "4" = "PR4", "5" = "PR5")
# RS_actual = c("1" = "RS1", "2" = "RS2")
# FR_actual = c("1" = "FR1", "2" = "FR2", "3" = "FR3")
# H_actual = c("0" = "0", "50" = "50", "100" = "100", "150" ="150", "200" = "200")
```


```{r, echo=FALSE, warning=FALSE}

np_actual = c("1" = expression(bolditalic(NP[1])), "2" = expression(bolditalic(NP[2])), "3"=expression(bolditalic(NP[3])))
fa_actual = c("1" = "NSF", "2" = "NIH", "3" = "DOE", "4" = "DOD", "5" = "NASA", "6" = "OT")
TS_actual = c("1" = expression(bolditalic(TS[1])), "2" = expression(bolditalic(TS[2])))
DS_actual = c("1" = expression(bolditalic(DS[1])), "2" = expression(bolditalic(DS[2])))
BF_actual = c("1" = expression(bolditalic(BF[1])), "2" = expression(bolditalic(BF[2])))
PR_actual = c("1" = expression(bolditalic(PR[1])), "2" = expression(bolditalic(PR[2])), "3" = expression(bolditalic(PR[3])), "4" = expression(bolditalic(PR[4])), "5" = expression(bolditalic(PR[5])))
RS_actual = c("1" = expression(bolditalic(RS[1])), "2" = expression(bolditalic(RS[2])))
FR_actual = c("1" = expression(bolditalic(FR[1])), "2" = expression(bolditalic(FR[2])), "3" = expression(bolditalic(FR[3])))
H_actual = c("0" = "0", "50" = "50", "100" = "100", "150" ="150", "200" = "200")
```


```{r, echo=FALSE, warning=FALSE}

get_legend_plot <-
  function(Data,
           item,
           modified_ticks,
           title,
           configuration, color) {
    
    axis_text_size <- 8
    # var_size = c('var1'=2, 'var2'=1)
    
    if (item == "NP" | item == "H" | item == "TA") {
      y_title = bquote( ~ italic("P"[.(paste0("G", configuration))]))
      y_text = element_text(face = "bold", size = axis_text_size)
    } else{
      y_title = element_blank()
      y_text = element_blank()
    }
    
    
    if (configuration != 20) {
      plot_title_ = element_blank()
      
    }
    if (configuration == 20) {
      plot_title_ = element_text(hjust = 0.5,
                                 face = "bold",
                                 size = axis_text_size)
    }
    
    
    Data <- filter(Data, Conf == configuration & Features == item)
    
    if (item == "NP" |
        item == "H" |
        item == "TA") {
      plot <- ggplot(Data, aes(x = Levels, y = fit_resp, colour = Levels)) +
        geom_errorbar(aes(ymin = Lower_limit, ymax = Upper_limit), size = 1.1) +
        theme_bw() +
        theme(
          axis.text.y = y_text,
          plot.title = plot_title_,
          panel.grid = element_blank(),
          axis.text.x = element_text(face = "bold", size = axis_text_size),
          legend.position = "bottom", 
          legend.direction = "horizontal",
          legend.title = element_blank(),
        #   legend.key.size = unit(2,"cm"),
        #   # legend.key.size = unit(1, 'cm'), #change legend key size
        # legend.key.height = unit(2, 'cm'), #change legend key height
        legend.key.width = unit(2, 'cm'),
          legend.text = element_text(size = 20)
        ) +
        labs(y = y_title, x = "", title = title) + scale_y_continuous(
          limits = c(0.0, 1),
          breaks = c(0.0, 0.25, 0.50, 0.75, 1),
          labels = c(
            "0.00" = "0%",
            "0.25" = "25%",
            "0.50" = "50%",
            "0.75" = "75%",
            "1" = "100%"
          )
        ) + theme(axis.ticks.y = element_blank(),
                  axis.ticks.x = element_blank()) +
        scale_x_discrete(labels = modified_ticks) +
        scale_color_manual(values = color, breaks=c("1", "2", "3"),
                       labels=c("*     ", "**     ", "***")) + coord_flip()
    }
  }


np_color = c("#4CAEE3", "#ee9a00", "red")
Legend_plot = get_legend_plot(Model_Data, "NP", np_actual, "title", 20, np_color)
# Legend_plot
legend <- get_legend(Legend_plot)

# 
 y_title = bquote( ~ italic("P"[.(29)]))
 # print(y_title)
```

```{r, echo=FALSE, warning=FALSE}
sectionA_plots <-
  function(Data,
           item,
           modified_ticks,
           title,
           configuration, color) {
    
    axis_text_size <- 8
    
    if (item =="FA"){
      font_face = "bold"  
    }else{
      font_face = "bold.italic"
    }
    
    if (item == "NP" | item == "H" | item == "TA") {
      # y_title = bquote( ~ italic("P"[.(paste0("G", configuration))]))
      y_title = bquote( ~ italic("S"^.(paste0("G", configuration))))
      y_text = element_text(face = "bold", size = axis_text_size)
    } else{
      y_title = element_blank()
      y_text = element_blank()
    }
    
    
    if (configuration != 20) {
      plot_title_ = element_blank()
      
    }
    if (configuration == 30) {
      plot_title_ = element_text(hjust = 0.5,
                                 face = "bold",
                                 size = axis_text_size)
    }
    
    
    Data <- filter(Data, Conf == configuration & Features == item)
    
    if (item == "NP" |
        item == "H" |
        item == "TA") {
      plot <- ggplot(Data, aes(x = Levels, y = fit_resp, colour = Levels)) +
        geom_pointrange(aes(ymin = Lower_limit, ymax = Upper_limit)) +
        theme_bw() +
        theme(
          axis.text.y = y_text,
          plot.title = plot_title_,
          panel.grid = element_blank(),
          axis.text.x = element_text(face = font_face, size = axis_text_size),
          # axis.text.y = element_text(face = "bold", size = 6),
          legend.position = "none"
        ) +
        labs(y = y_title, x = "", title = title) + scale_y_continuous(
          limits = c(0.0, 1),
          breaks = c(0.0, 0.25, 0.50, 0.75, 1),
          labels = c(
            "0.00" = "0%",
            "0.25" = "25%",
            "0.50" = "50%",
            "0.75" = "75%",
            "1" = "100%"
          )
        ) + theme(axis.ticks.y = element_blank(),
                  axis.ticks.x = element_blank()) +
        scale_x_discrete(labels = modified_ticks) +
        scale_color_manual(values = color)
      
      # print(plot)
      
    } else{
      plot <- ggplot(Data, aes(x = Levels, y = fit_resp, colour = Levels)) +
        geom_pointrange(aes(ymin = Lower_limit, ymax = Upper_limit)) +
        theme_bw() +
        theme(
          axis.text.y = y_text,
          plot.title = plot_title_,
          panel.grid = element_blank(),
          axis.text.x = element_text(face = font_face, size = axis_text_size),
          # axis.text.y = element_text(face = "bold", size = 6),
          legend.position = "none"
        ) +
        labs(y = y_title, x = "", title = title) + scale_y_continuous(
          limits = c(0.0, 1),
          breaks = c(0.0, 0.25, 0.50, 0.75, 1),
          labels = c(
            "0.00" = "0%",
            "0.25" = "25%",
            "0.50" = "50%",
            "0.75" = "75%",
            "1" = "100%"
          )
        ) + theme(axis.ticks.y = element_blank(),
                  axis.ticks.x = element_blank()) +
        scale_x_discrete(labels = modified_ticks) +
        scale_color_manual(values = color)
      # print(plot)
    }
  }
```



```{r, echo=FALSE, warning=FALSE}
continous_features_plots <-
  function(model_name,
           item,
           modified_ticks,
           Plot_title,
           configuration, clr) {
    axis_text_size <- 8
      # y_title = bquote(~ italic("P"[.(paste0("G", configuration))]))
      y_title = bquote(~ italic("S"^.(paste0("G", configuration))))
      
      y_text = element_text(face = "bold")
    
    
    
    if (configuration != "20") {
      plot_title_ = element_blank()
      
    }
    if (configuration == "30") {
      plot_title_ = element_text(hjust = 0.5,
                                 face = "bold",
                                 size = axis_text_size)
    }
    
    if (item == "H") {
      plt = plot_model(model_name, type = "pred", terms = item) +
        geom_line(aes(colour = item), size = 1) +
        scale_color_manual(values = c(clr)) +
        geom_vline(xintercept = 29.4, linetype="dashed", color = "gray", size=1)+
        theme_bw()+
        theme(
          legend.position = "none",
          plot.title = plot_title_,
          panel.grid = element_blank(),
          axis.text.x = element_text(face = "bold", size = axis_text_size),
          axis.text.y = element_text(face = "bold", size = axis_text_size)
        ) +
        labs(y = y_title, x = "", title = Plot_title) + scale_y_continuous(
          limits = c(0.0, 1),
          breaks = c(0.0, 0.25, 0.50, 0.75, 1),
          labels = c(
            "0.00" = "0%",
            "0.25" = "25%",
            "0.50" = "50%",
            "0.75" = "75%",
            "1" = "100%"
          )
        ) + theme(
          axis.ticks.y = element_blank(),
          axis.ticks.x = element_blank()
        )
      return(plt)
    } else if(item == "TA") {
      
      plt = plot_model(model_name, type = "pred", terms = item) +
        geom_line(aes(colour = item), size = 1) +
        scale_color_manual(values = c(clr)) +
        geom_vline(xintercept = 41.3, linetype="dashed", color = "gray", size=1)+
        theme_bw()+
        theme(
          legend.position = "none",
          plot.title = plot_title_,
          panel.grid = element_blank(),
          axis.text.x = element_text(face = "bold", size = axis_text_size),
          axis.text.y = element_blank()
        ) +
        labs(y = element_blank(), x = "", title = Plot_title) + scale_y_continuous(
          limits = c(0.0, 1),
          breaks = c(0.0, 0.25, 0.50, 0.75, 1),
          labels = c(
            "0.00" = "0%",
            "0.25" = "25%",
            "0.50" = "50%",
            "0.75" = "75%",
            "1" = "100%"
          )
        ) + theme(
          axis.ticks.y = element_blank(),
          axis.ticks.x = element_blank()
        )
      return(plt)
    } else if(item == "AC") {
      
      plt = plot_model(model_name, type = "pred", terms = item) +
        geom_line(aes(colour = item), size = 1) +
        scale_color_manual(values = c(clr)) +
        geom_vline(xintercept = 18.8, linetype="dashed", color = "gray", size=1)+
        theme_bw()+
        theme(
          legend.position = "none",
          plot.title = plot_title_,
          panel.grid = element_blank(),
          axis.text.x = element_text(face = "bold", size = axis_text_size),
          axis.text.y = element_blank()
        ) +
        labs(y = element_blank(), x = "", title = Plot_title) + scale_y_continuous(
          limits = c(0.0, 1),
          breaks = c(0.0, 0.25, 0.50, 0.75, 1),
          labels = c(
            "0.00" = "0%",
            "0.25" = "25%",
            "0.50" = "50%",
            "0.75" = "75%",
            "1" = "100%"
          )
        ) + theme(
          axis.ticks.y = element_blank(),
          axis.ticks.x = element_blank()
        )
      return(plt)
    } else {
      
      plt = plot_model(model_name, type = "pred", terms = item) +
        geom_line(aes(colour = item), size = 1) +
        scale_color_manual(values = c(clr)) +
        geom_vline(xintercept = 5.9, linetype="dashed", color = "gray", size=1)+
        theme_bw()+
        theme(
          legend.position = "none",
          plot.title = element_blank(),
          panel.grid = element_blank(),
          axis.text.x = element_text(face = "bold", size = axis_text_size),
          axis.text.y = element_blank()
        ) +
        labs(y = element_blank(), x = "", title = Plot_title) + scale_y_continuous(
          limits = c(0.0, 1),
          breaks = c(0.0, 0.25, 0.50, 0.75, 1),
          labels = c(
            "0.00" = "0%",
            "0.25" = "25%",
            "0.50" = "50%",
            "0.75" = "75%",
            "1" = "100%"
          )
        ) + theme(
          axis.ticks.y = element_blank(),
          axis.ticks.x = element_blank()
        )
      return(plt)
    }
    
  }
```



```{r, echo=FALSE, warning=FALSE}
dummy_df <- data.frame(
  x = c(0, 1, 0, 1),
  y = c(1, 0, 0, 1),
  group = c('a', 'a', 'b', 'b')
)

y_title = expression(paste(italic("P", "S20%")))

SectionA_draw_dummy_plot <- function(plot_title, y_title) {
  axis_text_size <- 8
  if (plot_title != "") {
    title = element_text(hjust = 0.5,
                         size = 8,
                         face = "bold")
  }
  
  if (plot_title == "") {
    title = element_blank()
  }
  dummy_plot <- ggplot(dummy_df, aes(x = x, y = y, group = group)) +
    geom_line(scale = 0.9, color = "white") +
    theme_bw() +
    theme(
      plot.title = title,
      axis.title.x = element_text(),
      axis.ticks.x = element_blank(),
      axis.ticks.y = element_blank(),
      axis.text.x = element_text(
        face = "bold",
        size = axis_text_size,
        angle = 30,
        hjust = 1
      ),
      axis.text.y = element_text(face = "bold", size = axis_text_size),
      panel.grid.major = element_blank(),
      panel.grid.minor = element_blank()
    ) + labs(y = element_blank(), x = "", title = plot_title) +
    scale_y_continuous(
      limits = c(0.0, 1),
      breaks = c(0.0, 0.25, 0.50, 0.75, 1),
      labels = c(
        "0.00" = "",
        "0.25" = "",
        "0.50" = "",
        "0.75" = "",
        "1" = ""
      )
    ) +
    scale_x_continuous(
      limits = c(0.0, 1),
      breaks = c(0.0, 0.25, 0.50, 0.75, 1),
      labels = c(
        "0.00" = "",
        "0.25" = "",
        "0.50" = "",
        "0.75" = "",
        "1" = ""
      )
    )
  
  return(dummy_plot)
  # dummy_plot
}




SectionB_draw_dummy_plot <-
  function(plot_title, y_title, configuration) {
    y_title = bquote( ~ italic("S"^.(paste0("G", configuration))))
    dummy_plot <- ggplot(dummy_df, aes(x = x, y = y, group = group)) +
      geom_line(scale = 0.9, color = "white") +
      theme_bw() +
      theme(
        plot.title = element_blank(),
        axis.title.x = element_text(),
        axis.ticks.x = element_blank(),
        axis.ticks.y = element_blank(),
        axis.text.x = element_text(
          face = "bold",
          size = 8,
          angle = 30,
          hjust = 1
        ),
        axis.text.y = element_text(face = "bold", size = 8),
        panel.grid.major = element_blank(),
        panel.grid.minor = element_blank()
      ) + labs(y = y_title, x = "", title = plot_title) +
      scale_y_continuous(
        limits = c(0.0, 1),
        breaks = c(0.0, 0.25, 0.50, 0.75, 1),
        labels = c(
          "0.00" = "0%",
          "0.25" = "25%",
          "0.50" = "50%",
          "0.75" = "75%",
          "1" = "100%"
        )
      ) +
      scale_x_continuous(
        limits = c(0.0, 1),
        breaks = c(0.0, 0.25, 0.50, 0.75, 1),
        labels = c(
          "0.00" = "",
          "0.25" = "",
          "0.50" = "",
          "0.75" = "",
          "1" = ""
        )
      )
    
    return(dummy_plot)
    dummy_plot
  }



SectionC_draw_dummy_plot <-
  function(plot_title, y_title, configuration) {
    if (plot_title != "") {
      title = element_text(hjust = 0.5,
                           size = 7,
                           face = "bold")
    }
    
    if (plot_title == "") {
      title = element_blank()
    }
    
    if (plot_title == "Trait Anxiety") {
      y_title = bquote(~ italic("S"^.(paste0("G", configuration))))
    }
    if (plot_title != "Trait Anxiety") {
      y_title = element_blank()
    }
    
    
    dummy_plot <-
      ggplot(dummy_df, aes(x = x, y = y, group = group)) +
      geom_line(scale = 0.9, color = "white") +
      theme_bw() +
      theme(
        plot.title = title,
        axis.title.x = element_text(),
        axis.ticks.y = element_blank(),
        axis.ticks.x = element_blank(),
        axis.text.x = element_text(
          face = "bold",
          size = 7,
          angle = 30,
          hjust = 1
        ),
        axis.text.y = element_text(face = "bold", size = 6),
        panel.grid.major = element_blank(),
        panel.grid.minor = element_blank()
      ) + labs(y = y_title, x = "", title = plot_title) +
      scale_y_continuous(
        limits = c(0.0, 1),
        breaks = c(0.0, 0.25, 0.50, 0.75, 1),
        labels = c(
          "0.00" = "0%",
          "0.25" = "25%",
          "0.50" = "50%",
          "0.75" = "75%",
          "1" = "100%"
        )
      ) + scale_x_continuous(
        limits = c(0.0, 1),
        breaks = c(0.0, 0.25, 0.50, 0.75, 1),
        labels = c(
          "0.00" = "",
          "0.25" = "",
          "0.50" = "",
          "0.75" = "",
          "1" = ""
        )
      )
    return(dummy_plot)
    # dummy_plot
  }

# draw_dummy_plot_blank_title <- function(plot_title) {
#   dummy_plot <- ggplot(dummy_df, aes(x=x, y=y, group=group)) + 
#     geom_line(scale = 0.9) +
#     # theme_minimal() +
#     theme(
#       plot.title = element_blank(),
#       axis.title.x=element_text(),
#         axis.text.x=element_blank(),
#         axis.ticks.x=element_blank(),
#         axis.title.y=element_blank(),
#         axis.text.y=element_blank(),
#         axis.ticks.y=element_blank(),
#         panel.grid.major = element_blank(),
#         panel.grid.minor = element_blank()) + labs(title =  plot_title, x= "")
#     
#   return(dummy_plot)
#   dummy_plot
# }

```


```{r, echo=FALSE, warning=FALSE}
orange = "#ee9a00"
blue = "#4CAEE3"

np_color = c("gray", "red", "red")
FA_color = c("gray", "black", "black", "red", "black", orange)
# ts_color = c("gray", orange)
color_list <- list(np_color, FA_color)


# plot_list <- list()
# plot_item <- c("NP", "FA")
# title <- c("Number of Proposals", "Funding Agency")
# 
# modified_ticx = list(np_actual, fa_actual, TS_actual, H_actual)
# color_list <- list(np_color, FA_color, ts_color)

# configuration = "20"
# for (item in 1:length(plot_item)){
#   plot = sectionA_plots(Model_Data, plot_item[item], modified_ticx[[item]], title[item], configuration, color_list[[item]])
#   plot_list[[length(plot_list) + 1]] <- plot
# }


# dummy_BF <- SectionA_draw_dummy_plot("Break Frequency", "")
# dummy_PR <- SectionA_draw_dummy_plot("Pilot Research", "")
# 
# 
# plot_A_20 <- ggarrange(
#   plot_list[[1]],
#   plot_list[[2]],
#   dummy_BF,
#   dummy_PR,
#   nrow = 1,
#   ncol = 4
# )
# plot_A_20



plot_list <- list()
plot_item <- c("NP", "FA")
title <- c("Number of Proposals", "Funding Agency")


np_color = c("gray", "red", "red")
FA_color = c("gray", "black", "black", orange, "black", orange)
# ts_color = c("gray", orange)
color_list <- list(np_color, FA_color)

modified_ticx = list(np_actual, fa_actual, FR_actual, RS_actual, H_actual , DS_actual)

configuration = "30"
for (item in 1:length(plot_item)){
  plot = sectionA_plots(Model_Data, plot_item[item], modified_ticx[[item]], title[item], configuration, color_list[[item]])
  plot_list[[length(plot_list) + 1]] <- plot
}


dummy_BF <- SectionA_draw_dummy_plot("Break Frequency", "")
dummy_PR <- SectionA_draw_dummy_plot("Pilot Research", "")


plot_A_30 <- ggarrange(
  plot_list[[1]],
  plot_list[[2]],
  # dummy_TS,
  dummy_BF,
  dummy_PR,
  nrow = 1,
  ncol = 4
)
# plot_A_30

 # + annotate(geom = 'text', x=.99, y=.5, label="S 30%", angle = 90, face = "bold")


plot_list <- list()
plot_item <- c("NP", "FA", "BF", "PR")
title <- c("", "", "", "")

np_color = c("gray", orange, "red")
FA_color = c("gray", "black", "black", blue, "black", orange)
BF_color = c("gray", blue)
PR_color = c("gray", blue, orange , blue, blue)
color_list <- list(np_color, FA_color, BF_color, PR_color)

modified_ticx = list(np_actual, fa_actual, BF_actual, PR_actual, H_actual, H_actual)
configuration = "50"
for (item in 1:length(plot_item)){
  plot = sectionA_plots(Model_Data, plot_item[item], modified_ticx[[item]], title[item], configuration, color_list[[item]])
  plot_list[[length(plot_list) + 1]] <- plot
}


dummy_TS <- SectionA_draw_dummy_plot("", "")


plot_A_50 <- ggarrange(
  plot_list[[1]],
  plot_list[[2]],
  plot_list[[3]],
  plot_list[[4]],
  nrow = 1,
  ncol = 4
)
# plot_A_50


# Research_tactics<- plot_grid(plot_A_20, plot_A_30, plot_A_50, nrow = 3)
Research_tactics<- ggarrange(plot_A_30, plot_A_50, nrow = 2)
# filename <- "sectionA.pdf"
# full_path <- file.path(plot_dir, filename)
# ggsave(
#   full_path,
#   Research_tactics,
#   width = 8.5,
#   height = 5.5,
#   units = "in"
# )

Research_tactics <- plot_grid(Research_tactics, legend, nrow = 2, rel_heights = c(1, .05))

```

```{r, echo=FALSE, warning=FALSE}

# plot1 = continous_features_plots(model_20, "H", H_actual, "h-index", 20)
plot2 = continous_features_plots(model_30, "H", H_actual, "h-index", 30, blue)
plot3 = continous_features_plots(model_50, "H", H_actual, "h-index", 50, orange)

sectionB <- ggarrange(
  # plot1,
  plot2,
  plot3,
  nrow = 2,
  ncol = 1
)
# sectionB


```

```{r, echo=FALSE, warning=FALSE}


plot_item <- c("DS")
title <- c("Deadline Stress")


DS_color = c("gray", blue)
color_list <- list(DS_color)

modified_ticx = list(DS_actual)

configuration = "20"
for (item in 1:length(plot_item)){
  # plotDS = sectionA_plots(Model_Data, plot_item[item], modified_ticx[[item]], title[item], configuration, color_list[[item]])
  plotDS30 = sectionA_plots(Model_Data, "DS", modified_ticx[[item]], title[item], 30, c("gray", orange))
}

# plot1 <- SectionA_draw_dummy_plot("Trait Anxiety", "")


plot2 <- SectionA_draw_dummy_plot("Extraversion", "")
plot3 <- SectionA_draw_dummy_plot("Avoidance coping", "")


plot4 <- continous_features_plots(model_30, "TA", FR_actual, "Trait Anxiety", 30, blue)
plot5 <- SectionA_draw_dummy_plot("", "")

plot6 <- SectionA_draw_dummy_plot("", "")
plot7 <- SectionA_draw_dummy_plot("", "")
plot8 = continous_features_plots(model_50, "E", H_actual, "", "", orange)
plot9 = continous_features_plots(model_50, "AC", H_actual, "", "", blue)

sectionC <- ggarrange(
  plotDS30,
  plot4,
  plot2,
  plot3,
  
  plot6,
  plot7,
  plot8,
  plot9,
  nrow = 2,
  ncol = 4
)

# sectionC <- ggarrange(
#   plotDS,
#   plot1,
#   plot2,
#   plot3,
#   plotDS30,
#   plot4,
#   plot5,
#   plot6,
#   plot5,
#   plot7,
#   plot8,
#   plot9,
#   nrow = 3,
#   ncol = 4
# )
# sectionC
```

```{r, echo=FALSE, warning=FALSE}

section_BC <- plot_grid(sectionB, sectionC, rel_widths = c(1, 2), ncol = 2, labels = c("b", "c"))

final_plot <-
  plot_grid(Research_tactics,
            section_BC,
            nrow = 2,
            labels = c("a", "")) + annotation_custom(grid.polygon(
              x = c(0, 0.5, 1, 0.336, 0.336),
              y = c(0.5, 0.5, 0.5, 0, 0.5),
              id = c(1, 1, 1, 2, 2),
              gp = gpar(lwd = 1.5))) +  annotation_custom(grid.polygon(
              # c(0, 0.25, 1, 1, 1),
              y = c(0.525, 0.525, 0.525, 0.525),
              id = NULL,
              gp = gpar(lwd = 1.5)))  






filename <- "Figure2.pdf"
full_path <- file.path(plot_dir, filename)
ggsave(
  full_path,
  final_plot,
  width = 8.5,
  height = 11,
  units = "in"
)
```
