==> scripts/7-vs-model-grant-funding.rmd <==
---
title: "Grant Funding Record Models"
output:
  html_document: default
  pdf_document: default
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(ggplot2)
library(stringr)
library(dplyr)
library(reshape2)
library(plyr)
library(aod)
library(pROC)
library(interplot)
library(sjPlot)
library(grid)
library(gridExtra)
library(ggpubr)
library(freqtables)
library(cowplot)
```

```{r, echo=FALSE}
root_dir <- getwd()
project_directory <- dirname(root_dir)
data_dir <- file.path(project_directory, 'raw-data')
curated_data_dir <- file.path(project_directory, 'curated-data')
plot_dir <- file.path(project_directory, 'Plots')


```


```{r, echo=FALSE}
data_file_name <- "KeyData.csv"
lm_DF <-
  read.csv(file.path(curated_data_dir, data_file_name),
           stringsAsFactors = FALSE)
lm_DF$SR <- NULL

# count(lm_DF$FC)
lm_DF <-
  lm_DF %>% dplyr::rename(TS = "T",
                          FR = "Rank",
                          E = "EXT",
                          PR = "AR")

lm_DF <-
  lm_DF %>%  mutate(FC = recode(
    FC,
    "1" = 0,
    "2" = 0,
    "3" = 1,
    "4" = 1,
    "5" = 1,
    "6" = 1
  ))

# count(lm_DF$FC)

lm_DF$RS <- as.factor(lm_DF$RS)
lm_DF$WH <- as.factor(lm_DF$WH)
lm_DF$NP <- as.factor(lm_DF$NP)
lm_DF$BF <- as.factor(lm_DF$BF)
lm_DF$DWH <- as.factor(lm_DF$DWH)
lm_DF$FA <- as.factor(lm_DF$FA)
lm_DF$TS <- as.factor(lm_DF$TS)
lm_DF$DS <- as.factor(lm_DF$DS)
lm_DF$PR <- as.factor(lm_DF$PR)
lm_DF$FR <- as.factor(lm_DF$FR)
lm_DF$FC <- as.factor(lm_DF$FC)

NullModel <- glm(FC ~ 1, data = lm_DF, family = "binomial")
FullModel <- glm(FC ~ ., data = lm_DF, family = "binomial")
BEmodel_25 <- stepAIC(FullModel, direction = "backward", trace = FALSE)
# summary(BEmodel_25)
```




```{r, echo=FALSE}
Zcrit<-1.96   ### critical value for 95% confidence intervals

ilink <- family(BEmodel_25)$linkinv
NP<-c("1","2","3",rep("1",12))
FA<-c(rep("1",3),"1","2","3","4","5","6",rep("1",6))
TS<-c(rep("1",9),"1","2",rep("1",4))
H<-rep(29.56824,15)
TA<-rep(41.33747,15)
OP<-rep(7.513648,15)
E<-rep(5.957816,15)
tmpd<-data.frame(NP,FA,TS,H,TA,OP,E)
ndata <- cbind(tmpd, setNames(as_tibble(predict(BEmodel_25, tmpd, se.fit = TRUE)[1:2]),
                                   c('fit_link','se_link')))
Model_Data_25 <- mutate(ndata,
                fit_resp  = ilink(fit_link),
                Lower_limit = ilink(fit_link - (Zcrit * se_link)),
                Upper_limit = ilink(fit_link + (Zcrit * se_link)))
# Model_Data_25
```


```{r, echo=FALSE}

data_file_name <- "KeyData.csv"
lm_DF <-
  read.csv(file.path(curated_data_dir, data_file_name),
           stringsAsFactors = FALSE)
lm_DF$SR <- NULL

# count(lm_DF$FC)
lm_DF <-
  lm_DF %>% dplyr::rename(TS = "T",
                          FR = "Rank",
                          E = "EXT",
                          PR = "AR")

lm_DF <-
  lm_DF %>%  mutate(FC = recode(
    FC,
    "1" = 0,
    "2" = 0,
    "3" = 0,
    "4" = 1,
    "5" = 1,
    "6" = 1
  ))

# count(lm_DF$FC)

lm_DF$RS <- as.factor(lm_DF$RS)
lm_DF$WH <- as.factor(lm_DF$WH)
lm_DF$NP <- as.factor(lm_DF$NP)
lm_DF$BF <- as.factor(lm_DF$BF)
lm_DF$DWH <- as.factor(lm_DF$DWH)
lm_DF$FA <- as.factor(lm_DF$FA)
lm_DF$TS <- as.factor(lm_DF$TS)
lm_DF$DS <- as.factor(lm_DF$DS)
lm_DF$PR <- as.factor(lm_DF$PR)
lm_DF$FR <- as.factor(lm_DF$FR)
lm_DF$FC <- as.factor(lm_DF$FC)


NullModel <- glm(FC ~ 1, data = lm_DF, family = "binomial")
FullModel <- glm(FC ~ ., data = lm_DF, family = "binomial")
BEmodel_50 <- stepAIC(FullModel, direction = "backward", trace = FALSE)


```


```{r, echo=FALSE}
Zcrit<-1.96   ### critical value for 95% confidence intervals

ilink <- family(BEmodel_50)$linkinv
NP<-c("1","2","3",rep("1",12))
FA<-c(rep("1",3),"1","2","3","4","5","6",rep("1",6))

H<-rep(29.56824,15)
TA<-rep(41.33747,15)
E<-rep(5.957816,15)


NASA<-rep(22.12903,15)
Task<-rep(28.44665,15)
TWR<-rep(42.08437,15)


tmpd<-data.frame(NP,FA,H,TA,E,NASA,Task,TWR)
ndata <- cbind(tmpd, setNames(as_tibble(predict(BEmodel_50, tmpd, se.fit = TRUE)[1:2]),
                                   c('fit_link','se_link')))
Model_Data_50 <- mutate(ndata,
                fit_resp  = ilink(fit_link),
                Lower_limit = ilink(fit_link - (Zcrit * se_link)),
                Upper_limit = ilink(fit_link + (Zcrit * se_link)))
# Model_Data_50
```





# Well Funded Record
```{r, echo=FALSE}

data_file_name <- "KeyData.csv"
lm_DF <-
  read.csv(file.path(curated_data_dir, data_file_name),
           stringsAsFactors = FALSE)
lm_DF$SR <- NULL

# count(lm_DF$FC)
lm_DF <-
  lm_DF %>% dplyr::rename(TS = "T",
                          FR = "Rank",
                          E = "EXT",
                          PR = "AR", 
                          O ="OP")

lm_DF <-
  lm_DF %>%  mutate(FC = recode(
    FC,
    "1" = 0,
    "2" = 0,
    "3" = 0,
    "4" = 0,
    "5" = 1,
    "6" = 1
  ))

# count(lm_DF$FC)

lm_DF$RS <- as.factor(lm_DF$RS)
lm_DF$WH <- as.factor(lm_DF$WH)
lm_DF$NP <- as.factor(lm_DF$NP)
lm_DF$BF <- as.factor(lm_DF$BF)
lm_DF$DWH <- as.factor(lm_DF$DWH)
lm_DF$FA <- as.factor(lm_DF$FA)
lm_DF$TS <- as.factor(lm_DF$TS)
lm_DF$DS <- as.factor(lm_DF$DS)
lm_DF$PR <- as.factor(lm_DF$PR)
lm_DF$FR <- as.factor(lm_DF$FR)
lm_DF$FC <- as.factor(lm_DF$FC)



NullModel <- glm(FC ~ 1, data = lm_DF, family = "binomial")
FullModel <- glm(FC ~ ., data = lm_DF, family = "binomial")
BEmodel_75 <- stepAIC(FullModel, direction = "backward", trace = FALSE)


```

```{r, echo=FALSE}

BEmodel_75 <- glm(formula = FC ~ O + H + RS + TWR + FA + TS, family = "binomial", data = lm_DF)
summary(BEmodel_75)

```


```{r, echo=FALSE}
Zcrit<-1.96   ### critical value for 95% confidence intervals
  
ilink <- family(BEmodel_75)$linkinv

FA<-c("1","2","3","4","5","6",rep("1",7))
RS<-c(rep("1",6),"1","2",rep("1",5))
TS<-c(rep("1",8),"1","2",rep("1",3))


# TA<-rep(41.33747,15)
# E<-rep(5.957816,15)
O<-rep(7.513648,13)
H<-rep(29.56824,13)
TWR<-rep(42.08437,13)


tmpd<-data.frame(FA,RS,TS,O,H,TWR)
ndata <- cbind(tmpd, setNames(as_tibble(predict(BEmodel_75, tmpd, se.fit = TRUE)[1:2]),
                                   c('fit_link','se_link')))
Model_Data_75 <- mutate(ndata,
                fit_resp  = ilink(fit_link),
                Lower_limit = ilink(fit_link - (Zcrit * se_link)),
                Upper_limit = ilink(fit_link + (Zcrit * se_link)))

```




\newpage
# Full Funded Record
```{r, echo=FALSE}

data_file_name <- "KeyData.csv"
lm_DF <-
  read.csv(file.path(curated_data_dir, data_file_name),
           stringsAsFactors = FALSE)
lm_DF$SR <- NULL
lm_DF$Disciplines <- NULL
lm_DF$State_Geo_location <- NULL

# count(lm_DF$FC)
lm_DF <-
  lm_DF %>% dplyr::rename(TS = "T",
                          FR = "Rank",
                          E = "EXT",
                          PR = "AR",
                          O = "OP")

lm_DF <-
  lm_DF %>%  mutate(FC = recode(
    FC,
    "1" = 0,
    "2" = 0,
    "3" = 0,
    "4" = 0,
    "5" = 0,
    "6" = 1
  ))

# count(lm_DF$FC)

lm_DF$RS <- as.factor(lm_DF$RS)
lm_DF$WH <- as.factor(lm_DF$WH)
lm_DF$NP <- as.factor(lm_DF$NP)
lm_DF$BF <- as.factor(lm_DF$BF)
lm_DF$DWH <- as.factor(lm_DF$DWH)
lm_DF$FA <- as.factor(lm_DF$FA)
lm_DF$TS <- as.factor(lm_DF$TS)
lm_DF$DS <- as.factor(lm_DF$DS)
lm_DF$PR <- as.factor(lm_DF$PR)
lm_DF$FR <- as.factor(lm_DF$FR)
lm_DF$FC <- as.factor(lm_DF$FC)

# mean(lm_DF$TWR)


NullModel <- glm(FC ~ 1, data = lm_DF, family = "binomial")
FullModel <- glm(FC ~ ., data = lm_DF, family = "binomial")
BEmodel_FF <- stepAIC(FullModel, direction = "backward", trace = FALSE)
# summary(BEmodel_FF)


BEmodel_FF <- glm(FC ~ O + TWR + FA, family = "binomial", data = lm_DF)
summary(BEmodel_FF)
```



```{r, echo=FALSE}
Zcrit<-1.96   ### critical value for 95% confidence intervals
  
ilink <- family(BEmodel_75)$linkinv

FA<-c("1","2","3","4","5","6",rep("1",2))
O<-rep(7.513648,8)
TWR<-rep(42.08437,8)


tmpd<-data.frame(FA,O,TWR)
ndata <- cbind(tmpd, setNames(as_tibble(predict(BEmodel_FF, tmpd, se.fit = TRUE)[1:2]),
                                   c('fit_link','se_link')))
Model_Data_FF <- mutate(ndata,
                fit_resp  = ilink(fit_link),
                Lower_limit = ilink(fit_link - (Zcrit * se_link)),
                Upper_limit = ilink(fit_link + (Zcrit * se_link)))
# Model_Data_75
```



\newpage
# Combined Grantsmanship and Grant Funding Record
```{r, echo=FALSE}

data_file_name <- "KeyData.csv"
lm_DF <-
  read.csv(file.path(curated_data_dir, data_file_name),
           stringsAsFactors = FALSE)
# lm_DF$SR <- NULL

# count(lm_DF$FC)
lm_DF <-
  lm_DF %>% dplyr::rename(TS = "T",
                          FR = "Rank",
                          E = "EXT",
                          PR = "AR",
                          O = "OP",
                          A = "AGR")

lm_DF <-
  lm_DF %>%  mutate(FC = recode(
    FC,
    "1" = 0,
    "2" = 0,
    "3" = 0,
    "4" = 0,
    "5" = 0,
    "6" = 1
  ))

# count(lm_DF$FC)

lm_DF$RS <- as.factor(lm_DF$RS)
lm_DF$WH <- as.factor(lm_DF$WH)
lm_DF$NP <- as.factor(lm_DF$NP)
lm_DF$BF <- as.factor(lm_DF$BF)
lm_DF$DWH <- as.factor(lm_DF$DWH)
lm_DF$FA <- as.factor(lm_DF$FA)
lm_DF$TS <- as.factor(lm_DF$TS)
lm_DF$DS <- as.factor(lm_DF$DS)
lm_DF$PR <- as.factor(lm_DF$PR)
lm_DF$FR <- as.factor(lm_DF$FR)
lm_DF$FC <- as.factor(lm_DF$FC)

# mean(lm_DF$TWR)



lm_DF <- lm_DF %>% mutate(Y5Zff = case_when(FC == "1" & SR > 4 ~ 1, TRUE ~ 0))

lm_DF$FC <- NULL
lm_DF$SR <- NULL

NullModel <- glm(Y5Zff ~ 1, data = lm_DF, family = "binomial")
FullModel <- glm(Y5Zff ~ ., data = lm_DF, family = "binomial")
BEmodel_FFSR50 <- stepAIC(FullModel, direction = "backward", trace= FALSE)
# summary(BEmodel_FFSR50)

BEmodel_FFSR50 <- glm(Y5Zff ~ E + A + TWR + NP, family = "binomial", data = lm_DF)
summary(BEmodel_FFSR50)
```

```{r, echo=FALSE}
Zcrit<-1.96   ### critical value for 95% confidence intervals
  
ilink <- family(BEmodel_FFSR50)$linkinv

NP<-c("1","2","3",rep("1",3))
E<-rep(5.957816,6)
A<-rep(7.528536,6)
TWR<-rep(42.08437,6)


tmpd<-data.frame(NP,E,A,TWR)
ndata <- cbind(tmpd, setNames(as_tibble(predict(BEmodel_FFSR50, tmpd, se.fit = TRUE)[1:2]),
                                   c('fit_link','se_link')))
Model_Data_FFS50 <- mutate(ndata,
                fit_resp  = ilink(fit_link),
                Lower_limit = ilink(fit_link - (Zcrit * se_link)),
                Upper_limit = ilink(fit_link + (Zcrit * se_link)))
```


```{r, echo=FALSE}
slice_df <- function(DF, R1, R2, Item, Conf){
  sliced_data = DF %>% 
        dplyr::slice(R1:R2) %>% 
        dplyr::select(Item, fit_resp, Lower_limit, Upper_limit) %>%
        dplyr::mutate(Features = Item) %>% 
        dplyr::mutate(Conf = Conf) %>% 
        dplyr::rename("Levels" = Item)
  return(sliced_data)
}
      
Model_Data <- slice_df(Model_Data_25, 1, 3, "NP", "25")

Model_Data <- rbind(Model_Data, slice_df(Model_Data_50, 1, 3, "NP", "50"))
Model_Data <- rbind(Model_Data, slice_df(Model_Data_50, 4, 9, "FA", "50"))

Model_Data <- rbind(Model_Data, slice_df(Model_Data_75, 1, 6, "FA", "75"))
Model_Data <- rbind(Model_Data, slice_df(Model_Data_75, 7, 8, "RS", "75"))
Model_Data <- rbind(Model_Data, slice_df(Model_Data_75, 9, 10, "TS", "75"))
Model_Data <- rbind(Model_Data, slice_df(Model_Data_FF, 1, 6, "FA", "FF"))
Model_Data <- rbind(Model_Data, slice_df(Model_Data_FFS50, 1, 3, "NP", "S50FF"))
```

```{r, echo=FALSE, warning=FALSE}

sectionA_plots <-
  function(Data,
           item,
           modified_ticks,
           title,
           configuration, color, suff) {
    
    axis_text_size <- 8
    

    
    if (item == "FA" | item == "H" | item == "NP") {
      y_title = bquote( ~ italic("S"^.(paste0(suff, configuration))))
      y_text = element_text(face = "bold", size = axis_text_size)
          if (configuration == "FF"){
            y_title = bquote( ~ italic("S"^.(paste0(suff, ""))))
      
    }
    } else{
      y_title = element_blank()
      y_text = element_blank()
    }
    
    
    if (configuration != 25) {
      plot_title_ = element_blank()
      
    }
    if (configuration == 75) {
      plot_title_ = element_text(hjust = 0.5,
                                 face = "bold",
                                 size = 7)
    }
    
    
    Data <- filter(Data, Conf == configuration & Features == item)
    
    if (item == "FA") {
      font_face = "bold"
    } else{
      font_face = "bold.italic"
    }
    
    if (item == "FA" |
        item == "H" |
        item == "NP") {
      plot <- ggplot(Data, aes(x = Levels, y = fit_resp, colour = Levels)) +
        geom_pointrange(aes(ymin = Lower_limit, ymax = Upper_limit)) +
        theme_bw() +
        theme(
          axis.text.y = y_text,
          plot.title = plot_title_,
          panel.grid = element_blank(),
          axis.text.x = element_text(face = "bold", size = axis_text_size),
          # axis.text.y = element_text(face = "bold", size = 6),
          legend.position = "none"
        ) +
        labs(y = y_title, x = "", title = title) + scale_y_continuous(
          limits = c(0.0, 1),
          breaks = c(0.0, 0.25, 0.50, 0.75, 1),
          labels = c(
            "0.00" = "0%",
            "0.25" = "25%",
            "0.50" = "50%",
            "0.75" = "75%",
            "1" = "100%"
          )
        ) + theme(axis.ticks.y = element_blank(),
                  axis.ticks.x = element_blank()) +
        scale_x_discrete(labels = modified_ticks) +
        scale_color_manual(values = color)
      
      # print(plot)
      
    } else{
      plot <- ggplot(Data, aes(x = Levels, y = fit_resp, colour = Levels)) +
        geom_pointrange(aes(ymin = Lower_limit, ymax = Upper_limit)) +
        theme_bw() +
        theme(
          axis.text.y = y_text,
          plot.title = plot_title_,
          panel.grid = element_blank(),
          axis.text.x = element_text(face = "bold", size = axis_text_size),
          # axis.text.y = element_text(face = "bold", size = 6),
          legend.position = "none"
        ) +
        labs(y = y_title, x = "", title = title) + scale_y_continuous(
          limits = c(0.0, 1),
          breaks = c(0.0, 0.25, 0.50, 0.75, 1),
          labels = c(
            "0.00" = "0%",
            "0.25" = "25%",
            "0.50" = "50%",
            "0.75" = "75%",
            "1" = "100%"
          )
        ) + theme(axis.ticks.y = element_blank(),
                  axis.ticks.x = element_blank()) +
        scale_x_discrete(labels = modified_ticks) +
        scale_color_manual(values = color)
      # print(plot)
    }
  }
```



```{r, echo=FALSE, warning=FALSE}
continous_features_plots <-
  function(model_name,
           item,
           modified_ticks,
           Plot_title,
           configuration,
           suff,
           clr) {
    
    axis_text_size <- 8
    y_title = bquote( ~ italic("S"^.(paste0(suff, configuration))))
    
    y_text = element_text(face = "bold")
    
    
    
    if (configuration != "20") {
      plot_title_ = element_blank()
      
    }
    if (configuration == "75" | configuration == "S50FF") {
      plot_title_ = element_text(hjust = 0.5,
                                 face = "bold",
                                 size = axis_text_size)
    }
    
    if (item == "H") {
      plt = plot_model(model_name, type = "pred", terms = item) +
        geom_line(aes(colour = item), size = 1) +
        scale_color_manual(values = c(clr)) +
        geom_vline(
          xintercept = 29.4,
          linetype = "dashed",
          color = "gray",
          size = 1
        ) +
        theme_bw() +
        theme(
          legend.position = "none",
          plot.title = plot_title_,
          panel.grid = element_blank(),
          axis.text.x = element_text(face = "bold", size = axis_text_size),
          axis.text.y = element_text(face = "bold", size = axis_text_size)
        ) +
        labs(y = y_title, x = "", title = Plot_title) + scale_y_continuous(
          limits = c(0.0, 1),
          breaks = c(0.0, 0.25, 0.50, 0.75, 1),
          labels = c(
            "0.00" = "0%",
            "0.25" = "25%",
            "0.50" = "50%",
            "0.75" = "75%",
            "1" = "100%"
          )
        ) + theme(axis.ticks.y = element_blank(),
                  axis.ticks.x = element_blank())
      return(plt)
    } else if (item == "TA") {
      plt = plot_model(model_name, type = "pred", terms = item) +
        geom_line(aes(colour = item), size = 1) +
        scale_color_manual(values = c(clr)) +
        geom_vline(
          xintercept = 41.3,
          linetype = "dashed",
          color = "gray",
          size = 1
        ) +
        theme_bw() +
        theme(
          legend.position = "none",
          plot.title = plot_title_,
          panel.grid = element_blank(),
          axis.text.x = element_text(face = "bold", size = axis_text_size),
          axis.text.y = element_blank()
        ) +
        labs(y = element_blank(), x = "", title = Plot_title) + scale_y_continuous(
          limits = c(0.0, 1),
          breaks = c(0.0, 0.25, 0.50, 0.75, 1),
          labels = c(
            "0.00" = "0%",
            "0.25" = "25%",
            "0.50" = "50%",
            "0.75" = "75%",
            "1" = "100%"
          )
        ) + theme(axis.ticks.y = element_blank(),
                  axis.ticks.x = element_blank())
      return(plt)
    } else if (item == "TWR") {
      plt = plot_model(model_name,
                       type = "pred",
                       terms = item) +
        geom_line(aes(colour = item), size = 1) +
        scale_color_manual(values = c(clr)) +
        geom_vline(
          xintercept = 42.08437,
          linetype = "dashed",
          color = "gray",
          size = 1
        ) +
        theme_bw() +
        theme(
          legend.position = "none",
          plot.title = plot_title_,
          panel.grid = element_blank(),
          axis.text.x = element_text(face = "bold", size = axis_text_size),
          axis.text.y = element_blank()
        ) +
        labs(y = element_blank(), x = "", title = Plot_title) + scale_y_continuous(
          limits = c(0.0, 1),
          breaks = c(0.0, 0.25, 0.50, 0.75, 1),
          labels = c(
            "0.00" = "0%",
            "0.25" = "25%",
            "0.50" = "50%",
            "0.75" = "75%",
            "1" = "100%"
          )
        ) + theme(axis.ticks.y = element_blank(),
                  axis.ticks.x = element_blank())
      return(plt)
    } else if (item == "E") {
      plt = plot_model(model_name, type = "pred", terms = item) +
        geom_line(aes(colour = item), size = 1) +
        scale_color_manual(values = c(clr)) +
        geom_vline(
          xintercept = 5.9,
          linetype = "dashed",
          color = "gray",
          size = 1
        ) +
        theme_bw() +
        theme(
          legend.position = "none",
          plot.title = plot_title_,
          panel.grid = element_blank(),
          axis.text.x = element_text(face = "bold", size = axis_text_size),
          axis.text.y = element_blank()
        ) +
        labs(y = element_blank(), x = "", title = Plot_title) + scale_y_continuous(
          limits = c(0.0, 1),
          breaks = c(0.0, 0.25, 0.50, 0.75, 1),
          labels = c(
            "0.00" = "0%",
            "0.25" = "25%",
            "0.50" = "50%",
            "0.75" = "75%",
            "1" = "100%"
          )
        ) + theme(axis.ticks.y = element_blank(),
                  axis.ticks.x = element_blank())
      return(plt)
    } else{
      plt = plot_model(model_name, type = "pred", terms = item) +
        geom_line(aes(colour = item), size = 1) +
        scale_color_manual(values = c(clr)) +
        geom_vline(
          xintercept = 7.5,
          linetype = "dashed",
          color = "gray",
          size = 1
        ) +
        theme_bw() +
        theme(
          legend.position = "none",
          plot.title = plot_title_,
          panel.grid = element_blank(),
          axis.text.x = element_text(face = "bold", size = axis_text_size),
          axis.text.y = element_blank()
        ) +
        labs(y = element_blank(), x = "", title = Plot_title) + scale_y_continuous(
          limits = c(0.0, 1),
          breaks = c(0.0, 0.25, 0.50, 0.75, 1),
          labels = c(
            "0.00" = "0%",
            "0.25" = "25%",
            "0.50" = "50%",
            "0.75" = "75%",
            "1" = "100%"
          )
        ) + theme(axis.ticks.y = element_blank(),
                  axis.ticks.x = element_blank())
      return(plt)
    }
    
  }
```


```{r, echo=FALSE, warning=FALSE}
dummy_df <- data.frame(
  x = c(0, 1, 0, 1),
  y = c(1, 0, 0, 1),
  group = c('a', 'a', 'b', 'b')
)

y_title = expression(paste(italic("P", "S20%")))

SectionA_draw_dummy_plot <- function(plot_title, y_title) {
  
  axis_text_size <- 8
  if (plot_title != "") {
    title = element_text(hjust = 0.5,
                         size = 7,
                         face = "bold")
  }
  
  if (plot_title == "") {
    title = element_blank()
  }
  dummy_plot <- ggplot(dummy_df, aes(x = x, y = y, group = group)) +
    geom_line(scale = 0.9, color = "white") +
    theme_bw() +
    theme(
      plot.title = title,
      axis.title.x = element_text(),
      axis.ticks.x = element_blank(),
      axis.ticks.y = element_blank(),
      axis.text.x = element_text(
        face = "bold",
        size = 6,
        angle = 30,
        hjust = 1
      ),
      axis.text.y = element_text(face = "bold", size = axis_text_size),
      panel.grid.major = element_blank(),
      panel.grid.minor = element_blank()
    ) + labs(y = element_blank(), x = "", title = plot_title) +
    scale_y_continuous(
      limits = c(0.0, 1),
      breaks = c(0.0, 0.25, 0.50, 0.75, 1),
      labels = c(
        "0.00" = "",
        "0.25" = "",
        "0.50" = "",
        "0.75" = "",
        "1" = ""
      )
    ) +
    scale_x_continuous(
      limits = c(0.0, 1),
      breaks = c(0.0, 0.25, 0.50, 0.75, 1),
      labels = c(
        "0.00" = "",
        "0.25" = "",
        "0.50" = "",
        "0.75" = "",
        "1" = ""
      )
    )
  
  return(dummy_plot)
  # dummy_plot
}




SectionB_draw_dummy_plot <-
  function(plot_title, y_title, configuration, suff) {
    
    axis_text_size <- 8
    y_title = bquote( ~ italic("S"^.(paste0(suff, configuration))))
    dummy_plot <- ggplot(dummy_df, aes(x = x, y = y, group = group)) +
      geom_line(scale = 0.9, color = "white") +
      theme_bw() +
      theme(
        plot.title = element_blank(),
        axis.title.x = element_text(),
        axis.ticks.x = element_blank(),
        axis.ticks.y = element_blank(),
        axis.text.x = element_text(
          face = "bold",
          size = 7,
          angle = 30,
          hjust = 1
        ),
        axis.text.y = element_text(face = "bold", size = axis_text_size),
        panel.grid.major = element_blank(),
        panel.grid.minor = element_blank()
      ) + labs(y = y_title, x = "", title = plot_title) +
      scale_y_continuous(
        limits = c(0.0, 1),
        breaks = c(0.0, 0.25, 0.50, 0.75, 1),
        labels = c(
          "0.00" = "0%",
          "0.25" = "25%",
          "0.50" = "50%",
          "0.75" = "75%",
          "1" = "100%"
        )
      ) +
      scale_x_continuous(
        limits = c(0.0, 1),
        breaks = c(0.0, 0.25, 0.50, 0.75, 1),
        labels = c(
          "0.00" = "",
          "0.25" = "",
          "0.50" = "",
          "0.75" = "",
          "1" = ""
        )
      )
    
    return(dummy_plot)
    # dummy_plot
  }



SectionC_draw_dummy_plot <-
  function(plot_title, y_title, configuration) {
    if (plot_title != "") {
      title = element_text(hjust = 0.5,
                           size = 7,
                           face = "bold")
    }
    
    if (plot_title == "") {
      title = element_blank()
    }
    
    if (plot_title == "Trait Anxiety") {
      y_title = bquote(~ italic("S"^.(configuration)))
    }
    if (plot_title != "Trait Anxiety") {
      y_title = element_blank()
    }
    
    
    dummy_plot <-
      ggplot(dummy_df, aes(x = x, y = y, group = group)) +
      geom_line(scale = 0.9, color = "white") +
      theme_bw() +
      theme(
        plot.title = title,
        axis.title.x = element_text(),
        axis.ticks.y = element_blank(),
        axis.ticks.x = element_blank(),
        axis.text.x = element_text(
          face = "bold",
          size = 7,
          angle = 30,
          hjust = 1
        ),
        axis.text.y = element_text(face = "bold", size = axis_text_size),
        panel.grid.major = element_blank(),
        panel.grid.minor = element_blank()
      ) + labs(y = y_title, x = "", title = plot_title) +
      scale_y_continuous(
        limits = c(0.0, 1),
        breaks = c(0.0, 0.25, 0.50, 0.75, 1),
        labels = c(
          "0.00" = "0%",
          "0.25" = "25%",
          "0.50" = "50%",
          "0.75" = "75%",
          "1" = "100%"
        )
      ) + scale_x_continuous(
        limits = c(0.0, 1),
        breaks = c(0.0, 0.25, 0.50, 0.75, 1),
        labels = c(
          "0.00" = "",
          "0.25" = "",
          "0.50" = "",
          "0.75" = "",
          "1" = ""
        )
      )
    return(dummy_plot)
    # dummy_plot
  }

```


```{r, echo=FALSE, warning=FALSE}
np_actual = c("1" = expression(bolditalic(NP[1])), "2" = expression(bolditalic(NP[2])), "3"=expression(bolditalic(NP[3])))
fa_actual = c("1" = "NSF", "2" = "NIH", "3" = "DOE", "4" = "DOD", "5" = "NASA", "6" = "OT")
TS_actual = c("1" = expression(bolditalic(TS[1])), "2" = expression(bolditalic(TS[2])))
DS_actual = c("1" = expression(bolditalic(DS[1])), "2" = expression(bolditalic(DS[2])))
BF_actual = c("1" = expression(bolditalic(BF[1])), "2" = expression(bolditalic(BF[2])))
PR_actual = c("1" = expression(bolditalic(PR[1])), "2" = expression(bolditalic(PR[2])), "3" = expression(bolditalic(PR[3])), "4" = expression(bolditalic(PR[4])), "5" = expression(bolditalic(PR[5])))
RS_actual = c("1" = expression(bolditalic(RS[1])), "2" = expression(bolditalic(RS[2])))
FR_actual = c("1" = expression(bolditalic(FR[1])), "2" = expression(bolditalic(FR[2])), "3" = expression(bolditalic(FR[3])))
H_actual = c("0" = "0", "50" = "50", "100" = "100", "150" ="150", "200" = "200")
```




```{r, echo=FALSE}
get_legend_plot <-
  function(Data,
           item,
           modified_ticks,
           title,
           configuration, color) {
    
    axis_text_size <- 8
    # var_size = c('var1'=2, 'var2'=1)
    
    if (item == "NP" | item == "H" | item == "TA") {
      y_title = bquote( ~ italic("P"[.(configuration)]))
      y_text = element_text(face = "bold", size = axis_text_size)
    } else{
      y_title = element_blank()
      y_text = element_blank()
    }
    
    
    if (configuration != 20) {
      plot_title_ = element_blank()
      
    }
    if (configuration == 25) {
      plot_title_ = element_text(hjust = 0.5,
                                 face = "bold",
                                 size = axis_text_size)
    }
    
    
    Data <- filter(Data, Conf == configuration & Features == item)
    
    if (item == "NP" |
        item == "H" |
        item == "TA") {
      plot <- ggplot(Data, aes(x = Levels, y = fit_resp, colour = Levels)) +
        geom_errorbar(aes(ymin = Lower_limit, ymax = Upper_limit), size = 1.1) +
        theme_bw() +
        theme(
          axis.text.y = y_text,
          plot.title = plot_title_,
          panel.grid = element_blank(),
          axis.text.x = element_text(face = "bold", size = axis_text_size),
          legend.position = "bottom", 
          legend.direction = "horizontal",
          legend.title = element_blank(),
        #   legend.key.size = unit(2,"cm"),
        #   # legend.key.size = unit(1, 'cm'), #change legend key size
        # legend.key.height = unit(2, 'cm'), #change legend key height
        legend.key.width = unit(2, 'cm'),
          legend.text = element_text(size = 20)
        ) +
        labs(y = y_title, x = "", title = title) + scale_y_continuous(
          limits = c(0.0, 1),
          breaks = c(0.0, 0.25, 0.50, 0.75, 1),
          labels = c(
            "0.00" = "0%",
            "0.25" = "25%",
            "0.50" = "50%",
            "0.75" = "75%",
            "1" = "100%"
          )
        ) + theme(axis.ticks.y = element_blank(),
                  axis.ticks.x = element_blank()) +
        scale_x_discrete(labels = modified_ticks) +
        scale_color_manual(values = color, breaks=c("1", "2", "3"),
                       labels=c("*     ", "**     ", "***")) + coord_flip()
    }
  }


np_color = c("#4CAEE3", "#ee9a00", "red")
Legend_plot = get_legend_plot(Model_Data, "NP", np_actual, "title", 25, np_color)
# Legend_plot
legend <- get_legend(Legend_plot)
```

```{r, echo=FALSE, warning=FALSE}
orange = "#ee9a00"
blue = "#4CAEE3"

# np_color = c("gray", blue, "red")
# FA_color = c("gray", "black", "black", "red", "black", orange)
# # ts_color = c("gray", orange)
# color_list <- list(np_color, FA_color)
# 
# 
# 
# modified_ticx = list(np_actual, fa_actual, TS_actual, H_actual)
# plot1 = sectionA_plots(Model_Data, "NP", np_actual, "Number of Proposals", "25", np_color)
# 
# 
# dummy_FA <- SectionA_draw_dummy_plot("Funding Agency", "")
# dummy_TS <- SectionA_draw_dummy_plot("Time of Submission", "")
# dummy_TWR <- SectionA_draw_dummy_plot("Typical Week Research", "")
# 
# 
# plot_A_25 <- ggarrange(
#   plot1,
#   dummy_FA,
#   dummy_TS,
#   dummy_TWR,
#   nrow = 1,
#   ncol = 4
# )
# # plot_A_25
# 
# 
# 
# plot_list <- list()
# plot_item <- c("NP", "FA")
# title <- c("", "")
# 
# 
# np_color = c("gray", blue, "red")
# FA_color = c("gray", "black", "black", blue, "black", "black")
# 
# color_list <- list(np_color, FA_color)
# 
# modified_ticx = list(np_actual, fa_actual, FR_actual, RS_actual, H_actual , DS_actual)
# 
# configuration = "50"
# for (item in 1:length(plot_item)){
#   plot = sectionA_plots(Model_Data, plot_item[item], modified_ticx[[item]], title[item], configuration, color_list[[item]])
#   plot_list[[length(plot_list) + 1]] <- plot
# }
# 
# 
# dummy_BF <- SectionA_draw_dummy_plot("", "")
# plot_twr = continous_features_plots(BEmodel_50, "TWR", H_actual, "", 50)
# 
# 
# plot_A_50 <- ggarrange(
#   plot_list[[1]],
#   plot_list[[2]],
#   dummy_BF,
#   plot_twr,
#   nrow = 1,
#   ncol = 4
# )

# plot_A_50



# dummy_NP <- SectionB_draw_dummy_plot("", "", "75")

plot_list <- list()
plot_item <- c("FA", "TS")
title <- c("Funding Agency", "Time of Submission")


FA_color = c("gray", "black", blue, orange, "black", "black")
TS_color = c("gray", blue)

color_list <- list(FA_color, TS_color)

modified_ticx = list(fa_actual, TS_actual)

configuration = "75"
for (item in 1:length(plot_item)){
  plot = sectionA_plots(Model_Data, plot_item[item], modified_ticx[[item]], title[item], configuration, color_list[[item]], "$")
  plot_list[[length(plot_list) + 1]] <- plot
}

plot_twr = continous_features_plots(BEmodel_75, "TWR", H_actual, "Typical Week Research", 75, "F", orange)


plot_A_75 <- ggarrange(
  plot_list[[1]],
  plot_list[[2]],
  plot_twr,
  nrow = 1,
  ncol = 3
)

# plot_A_75



plot_list <- list()
plot_item <- c("FA")
title <- c("Funding Agency")


FA_color = c("gray", "black", "black", "red", "black", "black")
# TS_color = c("gray", blue)

color_list <- list(FA_color)

modified_ticx = list(fa_actual)

configuration = "FF"
for (item in 1:length(plot_item)){
  plot = sectionA_plots(Model_Data, plot_item[item], modified_ticx[[item]], title[item], configuration, color_list[[item]], "$$")
  plot_list[[length(plot_list) + 1]] <- plot
}

dummy_TS <- SectionA_draw_dummy_plot("", "")
plot_twr = continous_features_plots(BEmodel_FF, "TWR", H_actual, "Typical Week Research", "FF", "", "red")


plot_A_FF <- ggarrange(
  plot_list[[1]],
  dummy_TS,
  plot_twr,
  nrow = 1,
  ncol = 3
)

# plot_A_FF



Research_tactics <- plot_grid(plot_A_75, plot_A_FF, nrow = 2)
# Research_tactics

Research_tactics <- plot_grid(Research_tactics, legend, nrow = 2, rel_heights = c(1, .05))
# Research_tactics
```



```{r, echo=FALSE, warning=FALSE}
plot5 = continous_features_plots(BEmodel_75, "H", H_actual, "h-index", 75, "$", blue)
RS_color = c("gray", blue)
plot6 = sectionA_plots(Model_Data, "RS", RS_actual, "Research Style", "75", RS_color, "F")
# plot7 = continous_features_plots(BEmodel_FF, "H", H_actual, "h-index", "FF")
plot7 <- SectionB_draw_dummy_plot("", "", "$$", "")
plot8 <- SectionA_draw_dummy_plot("", "")


sectionB <- plot_grid(plot5, plot6, plot7, plot8, nrow = 2, ncol = 2)
# sectionB
```



```{r, echo=FALSE, warning=FALSE}
plot1 <- SectionA_draw_dummy_plot("Openness", "")
plot2 = continous_features_plots(BEmodel_75, "O", H_actual, "Openness", "75", "F", blue)
plot3 = continous_features_plots(BEmodel_FF, "O", H_actual, "", "", "", blue)


sectionC <- plot_grid(plot2, plot3, nrow = 2, ncol = 1)
# sectionC
```


```{r, echo=FALSE, warning=FALSE}
section_BC <- plot_grid(sectionB, sectionC, rel_widths = c(2, 1), ncol = 2, labels = c("b", "c"))



final_plot <-
  plot_grid(Research_tactics,
            section_BC,
            nrow = 2,
            labels = c("a", "")) + annotation_custom(grid.polygon(
              x = c(0, 0.5, 1, 0.67, 0.67),
              y = c(0.5, 0.5, 0.5, 0, 0.5),
              id = c(1, 1, 1, 2, 2),
              gp = gpar(lwd = 1.5))) +  annotation_custom(grid.polygon(
              # c(0, 0.25, 1, 1, 1),
              y = c(0.525, 0.525, 0.525, 0.525),
              id = NULL,
              gp = gpar(lwd = 1.5)))



filename <- "Figure3_.pdf"
full_path <- file.path(plot_dir, filename)
ggsave(
  full_path,
  final_plot,
  width = 8.5,
  height = 11,
  units = "in"
)
```






```{r, echo=FALSE, warning=FALSE}
FFS50_plots <-
  function(Data,
           item,
           modified_ticks,
           title,
           configuration,
           color) {

    axis_text_size <- 8
    
    if (configuration != 25) {
      plot_title_ = element_blank()
      
    }
    if (configuration == "S50FF") {
      plot_title_ = element_text(hjust = 0.5,
                                 face = "bold",
                                 size = 7)
    }
    
    
    Data <- filter(Data, Conf == configuration & Features == item)
    
        if (item == "FA" | item == "H" | item == "NP") {
      y_title = bquote(~ italic("S"^.("S50$$")))
      y_text = element_text(face = "bold", size = axis_text_size)
    } else{
      y_title = element_blank()
      y_text = element_blank()
    }
    
    if (item == "FA" |
        item == "H" |
        item == "NP") {
      plot <-
        ggplot(Data, aes(x = Levels, y = fit_resp, colour = Levels)) +
        geom_pointrange(aes(ymin = Lower_limit, ymax = Upper_limit)) +
        theme_bw() +
        theme(
          axis.text.y = y_text,
          plot.title = plot_title_,
          panel.grid = element_blank(),
          axis.text.x = element_text(face = "bold", size = axis_text_size),
          # axis.text.y = element_text(face = "bold", size = 6),
          legend.position = "none"
        ) +
        labs(y = y_title, x = "", title = title) + scale_y_continuous(
          limits = c(0.0, 1),
          breaks = c(0.0, 0.25, 0.50, 0.75, 1),
          labels = c(
            "0.00" = "0%",
            "0.25" = "25%",
            "0.50" = "50%",
            "0.75" = "75%",
            "1" = "100%"
          )
        ) + theme(axis.ticks.y = element_blank(),
                  axis.ticks.x = element_blank()) +
        scale_x_discrete(labels = modified_ticks) +
        scale_color_manual(values = color)
      
      # print(plot)
      
    }
  }
```

```{r, echo=FALSE}
plot_item <- c("NP")
title <- c("Number of Proposals")


FA_color = c("gray",  orange, orange)
color_list <- list(FA_color)
modified_ticx = list(np_actual)

plot_S50_FF = FFS50_plots(Model_Data, plot_item[item], modified_ticx[[item]], title[item], "S50FF", color_list[[item]])
# plot_S50_FF

plot2 = continous_features_plots(BEmodel_FFSR50, "TWR", H_actual, "Typical Week Research", "S50FF", "", orange)
# plot2

plot3 = continous_features_plots(BEmodel_FFSR50, "E", H_actual, "Extraversion", "S50FF", "", blue)
# plot3

plot4 = continous_features_plots(BEmodel_FFSR50, "A", H_actual, "Agreeableness", "S50FF",  "", blue)
# plot4



final_plot <-
  plot_grid(plot_S50_FF,
            plot2,
            plot3,
            plot4,
            nrow = 1)



final_plot <- plot_grid(final_plot, legend, nrow = 2, rel_heights = c(1, .05))

filename <- "Figure4_.pdf"
full_path <- file.path(plot_dir, filename)
ggsave(
  full_path,
  final_plot,
  width = 8.5,
  height = 4,
  units = "in"
)

```
